findICF

Identifies cell-to-cell Interaction Changed Features (ICF), i.e. features that are differentially expressed due to proximity to other cell types.

findICF(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = "normalized",
  selected_feats = NULL,
  cluster_column,
  spatial_network_name = "Delaunay_network",
  minimum_unique_cells = 1,
  minimum_unique_int_cells = 1,
  diff_test = c("permutation", "limma", "t.test", "wilcox"),
  mean_method = c("arithmic", "geometric"),
  offset = 0.1,
  adjust_method = c("bonferroni", "BH", "holm", "hochberg", "hommel", "BY", "fdr",
    "none"),
  nr_permutations = 100,
  exclude_selected_cells_from_test = TRUE,
  do_parallel = TRUE,
  set_seed = TRUE,
  seed_number = 1234
)

Arguments

gobject

giotto object

feat_type

feature type

spat_unit

spatial unit

expression_values

expression values to use

selected_feats

subset of selected features (optional)

cluster_column

name of column to use for cell types

spatial_network_name

name of spatial network to use

minimum_unique_cells

minimum number of target cells required

minimum_unique_int_cells

minimum number of interacting cells required

diff_test

which differential expression test

mean_method

method to use to calculate the mean

offset

offset value to use when calculating log2 ratio

adjust_method

which method to adjust p-values

nr_permutations

number of permutations if diff_test = permutation

exclude_selected_cells_from_test

exclude interacting cells other cells

do_parallel

run calculations in parallel with mclapply

set_seed

set a seed for reproducibility

seed_number

seed number

Value

`icfObject` that contains the Interaction Changed differential gene scores

Details

Function to calculate if genes are differentially expressed in cell types when they interact (approximated by physical proximity) with other cell types. The results data.table in the `icfObject` contains - at least - the following columns:

features: All or selected list of tested features

sel: average feature expression in the interacting cells from the target cell type

other: average feature expression in the NOT-interacting cells from the target cell type

log2fc: log2 fold-change between sel and other

diff: spatial expression difference between sel and other

p.value: associated p-value

p.adj: adjusted p-value

cell_type: target cell type

int_cell_type: interacting cell type

nr_select: number of cells for selected target cell type

int_nr_select: number of cells for interacting cell type

nr_other: number of other cells of selected target cell type

int_nr_other: number of other cells for interacting cell type

unif_int: cell-cell interaction

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

findICF(g, cluster_column = "leiden_clus",
selected_feats = c("Gna12", "Ccnd2", "Btbd17"), nr_permutations = 10)
#> $ICFscores
#>       feats      sel     other     log2fc       diff p.value p.adj  perm_sel
#>      <char>    <num>     <num>      <num>      <num>   <num> <num>     <num>
#>   1:  Ccnd2 2.342484 3.0716941 -0.3769044 -0.7292099     0.1   0.3 2.4263862
#>   2:  Gna12 2.559092 3.5246024 -0.4468890 -0.9655102     0.1   0.3 2.6748786
#>   3: Btbd17 1.085768 1.5411328 -0.4688703 -0.4553648     0.1   0.3 1.1170828
#>   4:  Ccnd2 2.267062 1.8126603  0.3075171  0.4544021     0.0   0.0 2.0604929
#>   5:  Gna12 2.824606 4.1767141 -0.5482606 -1.3521083     0.0   0.0 3.4622022
#>  ---                                                                        
#> 119:  Ccnd2 2.998003 2.1417984  0.4666820  0.8562047     0.5   1.0 2.2731824
#> 120: Btbd17 0.000000 0.8799068 -3.2926445 -0.8799068     0.4   1.0 1.2330173
#> 121:  Gna12 3.550197 3.0520241  0.2116958  0.4981730     0.4   1.0 2.7958883
#> 122: Btbd17 0.000000 1.2064048 -3.7075301 -1.2064048     0.7   1.0 0.8588461
#> 123:  Ccnd2 0.000000 2.2244852 -4.5388393 -2.2244852     0.5   1.0 1.3870790
#>      perm_other  perm_log2fc    perm_diff cell_type int_cell_type nr_select
#>           <num>        <num>        <num>    <char>        <char>     <int>
#>   1:  2.2606417  0.142019659  0.165744471         3             3        87
#>   2:  2.4053338  0.253489729  0.269544755         3             3        87
#>   3:  1.2384223  0.005783867 -0.121339522         3             3        87
#>   4:  2.0581487  0.002120020  0.002344162         1             3        82
#>   5:  3.4189908  0.018261016  0.043211387         1             3        82
#>  ---                                                                       
#> 119:  2.1590560 -0.971157728  0.114126388         7             2         1
#> 120:  0.8505492 -0.622932823  0.382468067         7             2         1
#> 121:  3.0579635 -1.252477835 -0.262075258         2             7         1
#> 122:  1.1996422 -2.251336688 -0.340796145         2             7         1
#> 123:  2.2135633 -2.129593768 -0.826484309         2             7         1
#>      int_nr_select nr_other int_nr_other unif_int spec_int type_int
#>              <int>    <int>        <int>   <char>   <char>   <char>
#>   1:            87        9            9     3--3     3--3     homo
#>   2:            87        9            9     3--3     3--3     homo
#>   3:            87        9            9     3--3     3--3     homo
#>   4:            77       69           19     1--3     1--3   hetero
#>   5:            77       69           19     1--3     1--3   hetero
#>  ---                                                               
#> 119:             1       42          127     2--7     7--2   hetero
#> 120:             1       42          127     2--7     7--2   hetero
#> 121:             1      127           42     2--7     2--7   hetero
#> 122:             1      127           42     2--7     2--7   hetero
#> 123:             1      127           42     2--7     2--7   hetero
#> 
#> $Giotto_info
#> $Giotto_info$values
#> [1] "normalized"
#> 
#> $Giotto_info$cluster
#> [1] "leiden_clus"
#> 
#> $Giotto_info$`spatial network`
#> [1] "Delaunay_network"
#> 
#> 
#> $test_info
#> $test_info$test
#> [1] "permutation"
#> 
#> $test_info$p.adj
#> [1] "bonferroni"
#> 
#> $test_info$`min cells`
#> [1] 1
#> 
#> $test_info$`min interacting cells`
#> [1] 1
#> 
#> $test_info$`exclude selected cells`
#> [1] TRUE
#> 
#> $test_info$perm
#> [1] 10
#> 
#> 
#> attr(,"class")
#> [1] "icfObject" "list"