findMarkers_one_vs_all

Identify marker feats for all clusters in a one vs all manner.

findMarkers_one_vs_all(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  method = c("scran", "gini", "mast"),
  pval = 0.01,
  logFC = 0.5,
  min_feats = 10,
  min_genes = NULL,
  min_expr_gini_score = 0.5,
  min_det_gini_score = 0.5,
  detection_threshold = 0,
  rank_score = 1,
  adjust_columns = NULL,
  verbose = TRUE,
  ...
)

Arguments

gobject

giotto object

feat_type

feature type

spat_unit

spatial unit

expression_values

feat expression values to use

cluster_column

clusters to use

subset_clusters

selection of clusters to compare

method

method to use to detect differentially expressed feats

pval

scran & mast: filter on minimal p-value

logFC

scan & mast: filter on logFC

min_feats

minimum feats to keep per cluster, overrides pval and logFC

min_genes

deprecated, use min_feats

min_expr_gini_score

gini: filter on minimum gini coefficient for expression

min_det_gini_score

gini: filter minimum gini coefficient for detection

detection_threshold

gini: detection threshold for feat expression

rank_score

gini: rank scores to include

adjust_columns

mast: column in pDataDT to adjust for (e.g. detection rate)

verbose

be verbose

...

additional parameters for the findMarkers function in scran or zlm function in MAST

Value

data.table with marker feats

Details

Wrapper for all one vs all functions to detect marker feats for clusters.

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

findMarkers_one_vs_all(g, cluster_column = "leiden_clus")
#> using 'Scran' to detect marker feats. If used in published 
#>       research, please cite: Lun ATL, McCarthy DJ, Marioni JC (2016).
#>       'A step-by-step workflow for low-level analysis of single-cell RNA-seq 
#>       data with Bioconductor.'
#>       F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2. 
#> start with cluster  1start with cluster  2start with cluster  3start with cluster  4start with cluster  5start with cluster  6start with cluster  7start with cluster  8
#>       Top      p.value          FDR     logFC    feats cluster ranking
#>     <int>        <num>        <num>     <num>   <char>  <char>   <num>
#>  1:     5 1.205842e-02 0.9977859665 0.4874947   Bcl11b       1       1
#>  2:    10 3.511111e-02 0.9977859665 0.3806615     Lmo1       1       2
#>  3:    11 3.645417e-02 0.9977859665 0.3476473       F3       1       5
#>  4:    12 3.666126e-02 0.9977859665 0.3538903    Cnih3       1       4
#>  5:    13 3.731946e-02 0.9977859665 0.3190907  Ppp1r3c       1       7
#>  6:    18 5.572853e-02 0.9977859665 0.3293664    Rims2       1       6
#>  7:    21 5.953031e-02 0.9977859665 0.3791106     Gfap       1       3
#>  8:    32 8.486369e-02 0.9977859665 0.3016231     Gjc3       1      10
#>  9:    34 9.043051e-02 0.9977859665 0.3018724   Chrna4       1       9
#> 10:    96 2.028550e-01 0.9977859665 0.3136667    Prkcd       1       8
#> 11:     2 4.898493e-04 0.1552822127 0.6517265    Prr18       2       1
#> 12:     3 2.181421e-03 0.4610069551 0.5776650    Grb14       2       2
#> 13:     4 7.615942e-03 0.7450029491 0.4395064     Tprn       2       6
#> 14:     6 9.391527e-03 0.7450029491 0.3639458    Clic1       2      21
#> 15:     7 1.061062e-02 0.7450029491 0.4249516    Olig2       2       8
#> 16:     9 1.396410e-02 0.7450029491 0.4624166     Hrh3       2       3
#> 17:    11 1.514636e-02 0.7450029491 0.4313471   Tmbim1       2       7
#> 18:    13 1.666549e-02 0.7450029491 0.4180254  Carhsp1       2       9
#> 19:    15 2.109848e-02 0.7450029491 0.4410100  Tmem88b       2       5
#> 20:    17 2.203570e-02 0.7450029491 0.4477613    Ugt8a       2       4
#> 21:    41 6.164499e-02 0.8113396953 0.4105069    Tnnt1       2      10
#> 22:     4 4.743215e-03 0.4898416945 0.4554210   Arpp19       3      11
#> 23:     7 6.563966e-03 0.4898416945 0.6265833    Lamp5       3       1
#> 24:     9 9.164248e-03 0.4898416945 0.4195437   Galnt6       3      19
#> 25:    12 1.150519e-02 0.4898416945 0.4780622      Hlf       3       5
#> 26:    13 1.201764e-02 0.4898416945 0.5184660   Hs3st2       3       4
#> 27:    16 1.308128e-02 0.4898416945 0.6022550     Tbr1       3       2
#> 28:    18 1.556545e-02 0.5482497721 0.5446304     Myl4       3       3
#> 29:    22 1.996661e-02 0.5754015144 0.4711852     Cygb       3       8
#> 30:    23 2.289989e-02 0.6223004555 0.4779823    Ttc9b       3       6
#> 31:    35 3.435413e-02 0.6223004555 0.4751272   Ipcef1       3       7
#> 32:    47 5.698888e-02 0.7568991305 0.4674720   Mpped1       3      10
#> 33:    48 5.730467e-02 0.7568991305 0.4699471     Rprm       3       9
#> 34:    21 1.188335e-02 0.3410016493 0.6342812     Hopx       4       1
#> 35:    33 2.481303e-02 0.4510277350 0.4356998    Camkv       4       7
#> 36:    40 3.176643e-02 0.4510277350 0.5533260    Cldn5       4       2
#> 37:    42 3.237329e-02 0.4510277350 0.4343172     Vwa1       4       8
#> 38:    46 3.459463e-02 0.4510277350 0.5471637   Kcnip2       4       4
#> 39:    49 3.525905e-02 0.4510277350 0.5489968    Nptx1       4       3
#> 40:    54 4.307298e-02 0.5057086375 0.4464144    H2-D1       4       5
#> 41:    91 9.243808e-02 0.6440191315 0.4310444    Scn3b       4       9
#> 42:    95 1.042148e-01 0.6852362032 0.4297119     Wfs1       4      10
#> 43:   109 1.222588e-01 0.7087121227 0.4401159    Cabp7       4       6
#> 44:     5 9.245748e-03 0.7825743005 0.4615613    Enpp2       5       6
#> 45:    12 2.036044e-02 0.9920109750 0.5375139   Dock10       5       1
#> 46:    17 2.919840e-02 0.9920109750 0.4618758      Gss       5       5
#> 47:    18 4.052615e-02 0.9920109750 0.4479817    H2-K1       5       8
#> 48:    20 4.441037e-02 0.9920109750 0.4666989    Crip1       5       4
#> 49:    24 5.584673e-02 0.9920109750 0.5169952     Cygb       5       2
#> 50:    26 6.171130e-02 0.9920109750 0.4681081    Rasd1       5       3
#> 51:    28 6.385688e-02 0.9920109750 0.4133315   Homer3       5      10
#> 52:    31 7.067100e-02 0.9920109750 0.4488242  Serinc5       5       7
#> 53:    43 9.803715e-02 0.9920109750 0.4398918    C1ql2       5       9
#> 54:    10 3.390328e-03 0.1988387253 0.7578079     Nnat       6       1
#> 55:    26 2.169566e-02 0.5110768025 0.5813141    Itih3       6       4
#> 56:    36 3.052347e-02 0.5200416463 0.5759538    Acta2       6       5
#> 57:    41 3.433569e-02 0.5200416463 0.4930160    Pde8a       6       9
#> 58:    42 3.497909e-02 0.5200416463 0.6134005  Lrrc10b       6       2
#> 59:    48 4.451194e-02 0.5759263164 0.5246221     Rbp1       6       8
#> 60:    53 4.944463e-02 0.5759263164 0.5881766   Hs3st4       6       3
#> 61:    57 5.177886e-02 0.5759263164 0.5570237   Ifitm3       6       6
#> 62:    73 8.210912e-02 0.7131121173 0.5553475     Crym       6       7
#> 63:   137 1.847839e-01 0.8464299876 0.4337428    Cabp7       6      10
#> 64:    22 1.531032e-02 0.4412157056 0.7573867     Egr1       7       1
#> 65:    27 2.562918e-02 0.6018111752 0.6704896    Lrtm2       7       2
#> 66:    35 4.303690e-02 0.7612321738 0.6444753  Slc24a3       7       3
#> 67:    37 4.516865e-02 0.7739709220 0.5899327    Arl15       7       6
#> 68:    39 5.162405e-02 0.7910657401 0.5509159  Carhsp1       7       8
#> 69:    46 5.967867e-02 0.7910657401 0.6074286 Serpini1       7       4
#> 70:    47 6.104923e-02 0.7910657401 0.6012024   Vstm2a       7       5
#> 71:    54 8.892487e-02 0.9849130658 0.5837446   Pou3f1       7       7
#> 72:    57 1.023963e-01 0.9849130658 0.5426690    Itpka       7       9
#> 73:    69 1.222412e-01 0.9849130658 0.5312934   Fibcd1       7      10
#> 74:    87 9.457178e-05 0.0006891783 1.8693042    Evi2a       8       4
#> 75:    88 7.821288e-04 0.0056348828 2.1998577    Cabp7       8       2
#> 76:    89 1.582550e-03 0.0112734460 1.2678734 Serpini1       8      21
#> 77:    90 5.636351e-03 0.0397049615 1.3487616   Sowaha       8      14
#> 78:    97 2.754720e-02 0.1800507443 2.4042109   Tcf7l2       8       1
#> 79:    99 3.424917e-02 0.2193330717 1.8251830    Prkch       8       5
#> 80:   101 4.187901e-02 0.2628840864 1.8810982    Vat1l       8       3
#> 81:   102 5.242655e-02 0.3258669659 1.6304672     Sypl       8       6
#> 82:   104 6.424877e-02 0.3916703713 1.4815352    Timp3       8       9
#> 83:   114 8.391460e-02 0.4666829284 1.5160113     Gad2       8       7
#> 84:   125 1.197719e-01 0.6074829347 1.3994118    Prr18       8      10
#> 85:   132 1.341407e-01 0.6378055693 1.5129355  Zdhhc22       8       8
#>       Top      p.value          FDR     logFC    feats cluster ranking