findMastMarkers_one_vs_all

Identify marker feats for all clusters in a one vs all manner based on the MAST package.

findMastMarkers_one_vs_all(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  adjust_columns = NULL,
  pval = 0.001,
  logFC = 1,
  min_feats = 10,
  min_genes = NULL,
  verbose = TRUE,
  ...
)

Arguments

gobject

giotto object

feat_type

feature type

spat_unit

spatial unit

expression_values

feat expression values to use

cluster_column

clusters to use

subset_clusters

selection of clusters to compare

adjust_columns

column in pDataDT to adjust for (e.g. detection rate)

pval

filter on minimal p-value

logFC

filter on logFC

min_feats

minimum feats to keep per cluster, overrides pval and logFC

min_genes

deprecated, use min_feats

verbose

be verbose

...

additional parameters for the zlm function in MAST

Value

data.table with marker feats

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> a giotto python environment was found
#> Using python path:
#>  "/Users/yuanlab/Library/r-miniconda/envs/giotto_env/bin/pythonw"

findMastMarkers_one_vs_all(gobject = g, cluster_column = "leiden_clus")
#> using 'MAST' to detect marker feats. If used in published
#>         research, please cite: McDavid A, Finak G, Yajima M (2020).
#>         MAST: Model-based Analysis of Single Cell Transcriptomics.
#>         R package version 1.14.0, https://github.com/RGLab/MAST/.
#> start with cluster  1
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  2
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  3
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  4
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  5
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  6
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#> start with cluster  7
#> 
#> Done!
#> Combining coefficients and standard errors
#> Calculating log-fold changes
#> Calculating likelihood ratio tests
#> Refitting on reduced model...
#> 
#> Done!
#>        feats   Pr(>Chisq)      coef     ci.hi      ci.lo          fdr
#>       <char>        <num>     <num>     <num>      <num>        <num>
#>   1:     Ddn 4.742563e-48  2.324786  2.602948  2.0466234 3.006785e-45
#>   2:     Ttr 1.925765e-43  1.861501  2.100066  1.6229359 6.104676e-41
#>   3: Slc44a1 2.947434e-32 -1.697366 -1.349228 -2.0455033 6.228911e-30
#>   4:  Homer2 6.109638e-32  1.200966  1.575925  0.8260079 9.683776e-30
#>   5:   Cabp7 5.238447e-31  2.052926  2.440745  1.6651062 6.642351e-29
#>  ---                                                                 
#> 578:    Lcn2 7.557048e-05  1.064874  1.919267  0.2104803 6.654400e-04
#> 579:    Tns1 8.107545e-05  1.741728  2.519053  0.9644019 7.041348e-04
#> 580:   H2-K1 8.580490e-05  2.170945  2.894463  1.4474264 7.351393e-04
#> 581:    Rcn3 8.825585e-05  1.535251  2.364823  0.7056788 7.460561e-04
#> 582:    Bcam 1.258118e-04  1.997774  2.980463  1.0150846 9.847488e-04
#>          cluster ranking
#>           <char>   <int>
#>   1: 1_vs_others       1
#>   2: 1_vs_others       2
#>   3: 1_vs_others       3
#>   4: 1_vs_others       4
#>   5: 1_vs_others       5
#>  ---                    
#> 578: 7_vs_others      71
#> 579: 7_vs_others      73
#> 580: 7_vs_others      74
#> 581: 7_vs_others      75
#> 582: 7_vs_others      81