get10Xmatrix

This function creates an expression matrix from a 10X structured folder

get10Xmatrix(path_to_data, gene_column_index = 1)

Arguments

path_to_data
path to the 10X folder

gene_column_index
which column from the features or genes .tsv file to use for row ids

Value

sparse expression matrix from 10X

Details

A typical 10X folder is named raw_feature_bc_matrix or raw_feature_bc_matrix and tt has 3 files:

barcodes.tsv(.gz)

features.tsv(.gz) or genes.tsv(.gz)

matrix.mtx(.gz)

By default the first column of the features or genes .tsv file will be used, however if multiple annotations are provided (e.g. ensembl gene ids and gene symbols) the user can select another column.

Examples