getCellProximityGeneScores

Compute cell-cell interaction enrichment (observed vs expected)

getCellProximityGeneScores(
  gobject,
  spatial_network_name = "Delaunay_network",
  cluster_column = "louvain_clus.1",
  selected_genes = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  do_diff_test = TRUE,
  diff_test = c("t.test", "wilcox"),
  false_discovery_test = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
    "fdr", "none"),
  false_discovery_target = c("cell_interactions", "genes"),
  minimum_unique_cells = NA,
  fold_change_addendum = 0.1,
  in_two_directions = TRUE,
  exclude_selected_cells_from_test = F,
  do_parallel = TRUE,
  cores = NA,
  verbose = T
)

Arguments

gobject
giotto object

spatial_network_name
name of spatial network to use

cluster_column
name of column to use for clusters

selected_genes
selection of genes to perform calculations for

expression_values
expression values to use

do_diff_test
perform differential test

diff_test
which differential expression test

false_discovery_test
test to adjust p-values for multiple hypothesis testing

false_discovery_target
adjust p-values per cell-cell pair or per gene

minimum_unique_cells
minimum number of cells needed to proceed

fold_change_addendum
constant to add when calculating log2 fold-change

in_two_directions
shows enrichment in both directions: cell1-cell2, cell2-cell1

exclude_selected_cells_from_test
exclude certain cells from test

do_parallel
run enrichment calculations in parallel with mclapply

cores
number of cores to use if do_parallel = TRUE

verbose
verbose

Value

Cell Proximity Gene scores (CPGscores) in data.table format

Details

Function to calculate if genes are differentially expressed in cell types when they interact (according to physical proximity) with other cell types. The results data.table contains the following columns:

genes: All or selected list of tested genes

cell_expr_1: average gene expression in cell type 1 from unified_int cell-cell interaction

cell_expr_2: average gene expression in cell type 2 from unified_int cell-cell interaction

comb_expr: combined average gene expression in cell type 1 and 2 from unified_int cell-cell interaction

all_cell_expr_1: average gene expression for all cells from cell type 1

all_cell_expr_2: average gene expression for all cells from cell type 2

all_comb_expr: combined average gene expression for all cells from cell type 1 and 2

pval_1: p-value from test between interacting cells and all cells from cell type 1

pval_2: p-value from test between interacting cells and all cells from cell type 2

cell_type_1: first cell type of cell-cell interaction

cell_type_2: second cell type of cell-cell interaction

interaction: the cell-cell interaction, based on physical proximity

nr_1: number of cell type 1 in the unified cell-cell interaction

nr_2: number of cell type 2 in the unified cell-cell interaction

all_nr_1: number of all cell type 1 in the whole dataset

all_nr_2: number of all cell type 2 in the whole dataset

diff_spat: difference between comb_expr and all_comb_expr

diff_spat_1: difference between cell_expr_1 and all_cell_expr_1

diff_spat_2: difference between cell_expr_1 and all_cell_expr_1

log2fc_spat_1: fold-change of diff_spat_1

log2fc_spat_2: fold-change of diff_spat_2

log2fc_spat: fold-change of diff_spat

type_int: type of interaction

unified_int: interaction with alphabetically sorted cell type 1 and cell type 2

unif_int_rank: 1 or 2

fdr_1: fdr from test between interacting cells and all cells from cell type 1

fdr_2: fdr from test between interacting cells and all cells from cell type 2

Examples

    getCellProximityGeneScores(gobject)