getGeneToGeneScores

Compute gene-gene enrichment scores.

getGeneToGeneScores(
  CPGscore,
  selected_genes = NULL,
  specific_genes_1 = NULL,
  specific_genes_2 = NULL,
  min_cells = 5,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5,
  direction = c("both", "up", "down"),
  fold_change_addendum = 0.1,
  do_parallel = TRUE,
  cores = NA,
  verbose = TRUE
)

Arguments

CPGscore
CPGscore, output from getCellProximityGeneScores()

selected_genes
select subset of genes

specific_genes_1
specific source genes (see details)

specific_genes_2
specific target genes (see details)

min_cells
min number of cells threshold

min_spat_diff
spatial difference threshold

min_log2_fc
log2 fold-change threshold

direction
up or downregulation or both

fold_change_addendum
constant to add when calculating log2 fold-change

do_parallel
run enrichment calculations in parallel with mclapply

cores
number of cores to use if do_parallel = TRUE

verbose
verbose

min_pval
p-value threshold

Value

Gene to gene scores in data.table format

Details

This converts the single gene cell proximityscores into pairwise combinations of genes, which allows you to determine if 2 genes are differentially expressed in interacting cell types.

Examples

    getGeneToGeneScores(CPGscore)