makeSignMatrixRank

Function to convert a single-cell count matrix and a corresponding single-cell cluster vector into a rank matrix that can be used with the Rank enrichment option.

makeSignMatrixRank(
  sc_matrix,
  sc_cluster_ids,
  ties_method = c("random", "max"),
  gobject = NULL
)

Arguments

sc_matrix

matrix of single-cell RNAseq expression data

sc_cluster_ids

vector of cluster ids

ties_method

how to handle rank ties

gobject

if giotto object is given then only genes present in both datasets will be considered

Value

matrix

See also

Examples

sign_gene <- c(
    "Bcl11b", "Lmo1", "F3", "Cnih3", "Ppp1r3c", "Rims2", "Gfap",
    "Gjc3", "Chrna4", "Prkcd", "Prr18", "Grb14", "Tprn", "Clic1", "Olig2",
    "Hrh3", "Tmbim1", "Carhsp1", "Tmem88b", "Ugt8a", "Arpp19", "Lamp5",
    "Galnt6", "Hlf", "Hs3st2", "Tbr1", "Myl4", "Cygb", "Ttc9b", "Ipcef1"
)

sign_matrix <- matrix(rnorm(length(sign_gene) * 3), nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- c("cell_type1", "cell_type2", "cell_type3")

makeSignMatrixRank(
    sc_matrix = sign_matrix,
    sc_cluster_ids = c("cell_type1", "cell_type2", "cell_type3")
)
#> Warning: NaNs produced
#> Warning: NaNs produced
#> 30 x 3 Matrix of class "dgeMatrix"
#>         cell_type1 cell_type2 cell_type3
#> Bcl11b          26         13          2
#> Lmo1            29         25         28
#> F3              20         24          8
#> Cnih3           23         22          6
#> Ppp1r3c         21          6         22
#> Rims2            8          2         29
#> Gfap             3         27          3
#> Gjc3             7         18         17
#> Chrna4          24          4          5
#> Prkcd           25          7         14
#> Prr18           17          9         20
#> Grb14           12          1         23
#> Tprn            27         19          4
#> Clic1           13         10         19
#> Olig2           11         16         15
#> Hrh3            19         21         10
#> Tmbim1          28         30         24
#> Carhsp1          5         26          1
#> Tmem88b          6         15         21
#> Ugt8a           30          8          9
#> Arpp19          10         17         16
#> Lamp5            1         29         25
#> Galnt6           4         12         26
#> Hlf             14         28          7
#> Hs3st2          15         11         18
#> Tbr1            16         23         11
#> Myl4             2          3         27
#> Cygb            22         14         13
#> Ttc9b            9          5         30
#> Ipcef1          18         20         12