Run dimension reduction method

Wrapper function for Giotto dimension reduction methods for easier coding.

reduceDims(
  gobject,
  method = c("pca", "nmf", "umap", "tsne"),
  projection = FALSE,
  toplevel = 1L,
  ...
)

Arguments

gobject

giotto object

method

character. Dimension reduction method to use

projection

logical. Whether to run in a projection manner (faster, but is an approximation)

toplevel

relative stackframe the call was made at. do not use.

...

additional params to pass to specific functions

Value

`giotto` object with attached dimension reduction

Examples

g <- GiottoData::loadGiottoMini("vis")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> python already initialized in this session
#>  active environment : '/usr/bin/python3'
#>  python version : 3.10
#> checking default envname 'giotto_env'
#> a system default python environment was found
#> Using python path:
#>  "/usr/bin/python3"
x <- reduceDims(g, "tsne", spat_unit = "cell")
#> > tsne already exists and will be replaced with
#>  new dimension reduction object
#> Setting dimension reduction [cell][rna] tsne
x <- reduceDims(x, "umap", projection = TRUE)
#> Setting dimension reduction [cell][rna] umap.projection
x <- reduceDims(x, method = "nmf")
#> "hvf" column was found in the feats metadata information and will be
#>  used to select highly variable features
#> [Running RcppML NMF]. This citation shown once per session:
#> Zachary J. DeBruine, Karsten Melcher, Timothy J. Triche Jr
#>  Fast and robust non-negative matrix factorization for single-cell experiments
#>  bioRxiv 2021.09.01.458620.
#>  https://doi.org/10.1101/2021.09.01.458620
#> 
#> iter |      tol 
#> ---------------
#>    1 | 8.68e-01
#>    2 | 1.06e-01
#>    3 | 4.20e-02
#>    4 | 2.40e-02
#>    5 | 1.50e-02
#>    6 | 1.07e-02
#>    7 | 8.13e-03
#>    8 | 5.39e-03
#>    9 | 3.11e-03
#>   10 | 2.04e-03
#>   11 | 1.53e-03
#>   12 | 1.18e-03
#>   13 | 9.46e-04
#>   14 | 7.65e-04
#>   15 | 6.38e-04
#>   16 | 5.10e-04
#>   17 | 4.68e-04
#>   18 | 4.30e-04
#>   19 | 3.94e-04
#>   20 | 3.51e-04
#>   21 | 3.07e-04
#>   22 | 2.39e-04
#>   23 | 1.90e-04
#>   24 | 1.70e-04
#>   25 | 1.53e-04
#>   26 | 1.35e-04
#>   27 | 1.21e-04
#>   28 | 1.05e-04
#>   29 | 9.11e-05
#> Setting dimension reduction [cell][rna] nmf