signPCA

identify significant prinicipal components (PCs)

signPCA(
  gobject,
  method = c("screeplot", "jackstraw"),
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  genes_to_use = NULL,
  scale_unit = T,
  ncp = 50,
  scree_labels = T,
  scree_ylim = c(0, 10),
  jack_iter = 10,
  jack_threshold = 0.01,
  jack_verbose = T,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "signPCA",
  ...
)

Arguments

gobject
giotto object

method
method to use to identify significant PCs

expression_values
expression values to use

reduction
cells or genes

genes_to_use
subset of genes to use for PCA

scale_unit
scale features before PCA

ncp
number of principal components to calculate

scree_labels
show labels on scree plot

scree_ylim
y-axis limits on scree plot

jack_iter
number of interations for jackstraw

jack_threshold
p-value threshold to call a PC significant

jack_verbose
show progress of jackstraw method

show_plot
show plot

return_plot
return ggplot object

save_plot
directly save the plot [boolean]

save_param
list of saving parameters from all_plots_save_function()

default_save_name
default save name for saving, don't change, change save_name in save_param

...
additional parameters for PCA

Value

ggplot object for scree method and maxtrix of p-values for jackstraw

Details

Two different methods can be used to assess the number of relevant or significant prinicipal components (PC's).
1. Screeplot works by plotting the explained variance of each individual PC in a barplot allowing you to identify which PC does not show a significant contribution anymore ( = 'elbow method').
2. The Jackstraw method uses the permutationPA function. By systematically permuting genes it identifies robust, and thus significant, PCs.
multiple PCA results can be stored by changing the name parameter

Examples

    signPCA(gobject)