spark

Compute spatially expressed genes with SPARK method

spark(
  gobject,
  spat_loc_name = "raw",
  feat_type = NULL,
  spat_unit = NULL,
  percentage = 0.1,
  min_count = 10,
  expression_values = "raw",
  num_core = 5,
  covariates = NULL,
  return_object = c("data.table", "spark"),
  ...
)

Arguments

gobject

giotto object

spat_loc_name

name for spatial locations

feat_type

feature type

spat_unit

spatial unit

percentage

The percentage of cells that are expressed for analysis

min_count

minimum number of counts for a gene to be included

expression_values

type of values to use (raw by default)

num_core

number of cores to use

covariates

The covariates in experiments, i.e. confounding factors/batch effect. Column name of giotto cell metadata.

return_object

type of result to return (data.table or spark object)

...

Additional parameters to the spark.vc function

Value

data.table with SPARK spatial genes results or the SPARK object

Details

This function is a wrapper for the method implemented in the SPARK package: SPARK package:

CreateSPARKObject create a SPARK object from a giotto object

spark.vc Fits the count-based spatial model to estimate the parameters, see spark.vc for additional parameters

spark.test Testing multiple kernel matrices

Publication: doi:10.1101/810903

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> a giotto python environment was found
#> Using python path:
#>  "/Users/yuanlab/Library/r-miniconda/envs/giotto_env/bin/pythonw"

spark(g)
#> using 'SPARK' for spatial gene/pattern detection. If used in
#>     published research, please cite:
#>     Sun, Shiquan, Jiaqiang Zhu, and Xiang Zhou. 'Statistical Analysis of
#>     Spatial Expression Pattern for Spatially Resolved Transcriptomic Studies.'
#>     BioRxiv, October 21, 2019, 810903. https://doi.org/10.1101/810903.
#> ## ===== SPARK INPUT INFORMATION ==== 
#> ## number of total samples:  624 
#> ## number of total features:  576 
#> ## number of adjusted covariates:  0 
#> # fitting count-based spatial model under the null hypothesis ... 
#> ## testing Gaussian kernel: 1...
#> ## testing Periodic kernel: 1...
#> ## testing Gaussian kernel: 2...
#> ## testing Periodic kernel: 2...
#> ## testing Gaussian kernel: 3...
#> ## testing Periodic kernel: 3...
#> ## testing Gaussian kernel: 4...
#> ## testing Periodic kernel: 4...
#> ## testing Gaussian kernel: 5...
#> ## testing Periodic kernel: 5...
#>              GSP1      COS1         GSP2      COS2         GSP3       COS3
#>             <num>     <num>        <num>     <num>        <num>      <num>
#>   1: 1.602064e-29 0.9626485 8.428986e-29 0.8746052 1.650841e-22 0.11333318
#>   2: 7.700079e-34 0.9996967 1.678228e-30 0.9999161 7.258092e-27 0.01100897
#>   3: 1.657064e-57 0.9995031 2.002956e-64 0.9699135 1.297105e-58 0.71870162
#>   4: 3.520553e-21 0.8601842 3.292507e-25 0.8338627 6.482156e-26 0.16420425
#>   5: 3.652749e-42 0.9999978 6.220618e-42 0.9904633 4.684591e-33 0.87747870
#>  ---                                                                      
#> 572: 1.647226e-01 0.6121844 2.277559e-01 0.7429497 5.709877e-01 0.76957544
#> 573: 5.982327e-01 0.8942523 3.619736e-01 0.5629510 3.324191e-01 0.74180068
#> 574: 2.650141e-01 0.3496706 1.453781e-01 0.9886919 6.466572e-02 0.87815719
#> 575: 2.112557e-01 0.5877784 5.979924e-01 0.9856482 8.825371e-01 0.75095905
#> 576: 6.732644e-01 0.8912192 8.015104e-01 0.1297482 8.958441e-01 0.99644824
#>              GSP4         COS4         GSP5         COS5 combined_pvalue
#>             <num>        <num>        <num>        <num>           <num>
#>   1: 9.896061e-18 1.652221e-02 4.192463e-14 1.827476e-11    5.550000e-17
#>   2: 2.071635e-22 3.492188e-05 6.823466e-18 8.573930e-04    5.550000e-17
#>   3: 2.542640e-43 1.719199e-03 2.261825e-28 1.784737e-06    5.550000e-17
#>   4: 1.569419e-25 3.102681e-02 4.541649e-23 1.387068e-08    5.550000e-17
#>   5: 8.508284e-22 1.298126e-02 8.916407e-14 1.491978e-05    5.550000e-17
#>  ---                                                                    
#> 572: 7.495441e-01 9.421084e-01 8.107784e-01 4.899843e-02    5.394030e-01
#> 573: 5.443894e-01 5.845903e-01 7.475274e-01 7.890190e-01    6.720542e-01
#> 574: 4.483906e-02 4.308208e-01 4.475741e-02 4.422972e-01    7.106458e-01
#> 575: 8.667364e-01 1.935389e-01 7.959211e-01 2.791300e-01    8.847304e-01
#> 576: 8.533334e-01 4.233712e-01 7.451522e-01 9.380562e-01    9.691529e-01
#>      adjusted_pvalue   genes
#>                <num>  <char>
#>   1:    8.761748e-16   Ccnd2
#>   2:    8.761748e-16    Sez6
#>   3:    8.761748e-16   Tiam1
#>   4:    8.761748e-16   Rab3b
#>   5:    8.761748e-16  Homer3
#>  ---                        
#> 572:    1.000000e+00   Paqr5
#> 573:    1.000000e+00 Unc93b1
#> 574:    1.000000e+00   Sox18
#> 575:    1.000000e+00  Pdgfra
#> 576:    1.000000e+00   Itpr2