Split cluster annotations based on a spatial network

Split cluster annotations based on a spatial network

spatialSplitCluster(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "Delaunay_network",
  cluster_col,
  split_clus_name = paste0(cluster_col, "_split"),
  include_all_ids = TRUE,
  missing_id_name = "not_connected",
  return_gobject = TRUE
)

Arguments

gobject

giotto object

spat_unit

spatial unit (e.g. "cell")

feat_type

feature type (e.g. "rna", "dna", "protein")

spatial_network_name

character. Name of spatial network to use

cluster_col

character. Column in metadata containing original clustering

split_clus_name

character. Name to assign the split cluster results

include_all_ids

Boolean. Include all ids, including vertex ids not found in the spatial network

missing_id_name

Character. Name for vertices that were missing from spatial network

return_gobject

Boolean. Return giotto object

Value

giotto object with cluster annotations

Examples

library(Giotto)
g <- GiottoData::loadGiottoMini("vizgen")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> Error in .os_py_path(must_exist = TRUE): Assertion on 'path' failed: Directory '/home/runner/.local/share/r-miniconda' does not exist.
activeSpatUnit(g) <- "aggregate"
#> Error: object 'g' not found
spatPlot2D(g, cell_color = "leiden_clus")
#> Error in eval(expr, envir, enclos): object 'g' not found

g <- spatialSplitCluster(g,
    cluster_col = "leiden_clus",
    split_clus_name = "new"
)
#> Warning: restarting interrupted promise evaluation
#> Warning: restarting interrupted promise evaluation
#> Error in eval(expr, envir, enclos): object 'g' not found
# don't show legend since there are too many categories generated
spatPlot2D(g, cell_color = "new", show_legend = FALSE)
#> Error in eval(expr, envir, enclos): object 'g' not found