cellProximityEnrichment

Compute cell-cell interaction enrichment (observed vs expected)

cellProximityEnrichment(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "Delaunay_network",
  cluster_column,
  number_of_simulations = 1000,
  adjust_method = c("none", "fdr", "bonferroni", "BH", "holm", "hochberg", "hommel",
    "BY"),
  set_seed = TRUE,
  seed_number = 1234
)

Arguments

gobject

giotto object

spat_unit

spatial unit

feat_type

feature type

spatial_network_name

name of spatial network to use

cluster_column

name of column to use for clusters

number_of_simulations

number of simulations to create expected observations

adjust_method

method to adjust p.values

set_seed

use of seed

seed_number

seed number to use

Value

List of cell Proximity scores (CPscores) in data.table format. The first data.table (raw_sim_table) shows the raw observations of both the original and simulated networks. The second data.table (enrichm_res) shows the enrichment results.

Details

Spatial proximity enrichment or depletion between pairs of cell types is calculated by calculating the observed over the expected frequency of cell-cell proximity interactions. The expected frequency is the average frequency calculated from a number of spatial network simulations. Each individual simulation is obtained by reshuffling the cell type labels of each node (cell) in the spatial network.

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

cellProximityEnrichment(g, cluster_column = "leiden_clus")
#> $raw_sim_table
#> Index: <unified_int>
#>        unified_int type_int    round    V1     orig
#>             <char>   <char>   <char> <num>   <char>
#>     1:        3--3     homo original   128 original
#>     2:        1--3   hetero original   168 original
#>     3:        3--6   hetero original    32 original
#>     4:        1--2   hetero original    83 original
#>     5:        2--2     homo original   297 original
#>    ---                                             
#> 46022:        5--8   hetero   sim927     0 original
#> 46023:        5--8   hetero   sim930     0 original
#> 46024:        6--8   hetero   sim950     0 original
#> 46025:        7--8   hetero   sim961     0 original
#> 46026:        4--8   hetero   sim963     0 original
#> 
#> $enrichm_res
#>     unified_int type_int original simulations     enrichm p_higher_orig
#>          <fctr>   <char>    <num>       <num>       <num>         <num>
#>  1:        7--7     homo       65       9.094  2.70896803         0.000
#>  2:        6--6     homo      105      16.878  2.56780701         0.000
#>  3:        5--5     homo      121      20.678  2.49257758         0.000
#>  4:        4--4     homo      107      23.059  2.16638273         0.000
#>  5:        2--2     homo      297      67.846  2.11386759         0.000
#>  6:        8--8     homo        3       0.168  1.77595973         0.000
#>  7:        4--5   hetero      137      43.688  1.62670888         0.000
#>  8:        3--3     homo      128      45.954  1.45805109         0.000
#>  9:        1--8   hetero       23       8.470  1.34159808         0.000
#> 10:        3--7   hetero       99      40.568  1.26645476         0.000
#> 11:        1--1     homo      224     102.937  1.11421568         0.000
#> 12:        1--3   hetero      168     137.952  0.28243665         0.001
#> 13:        1--6   hetero       80      83.394 -0.05921853         0.692
#> 14:        2--8   hetero        5       6.623 -0.34539638         0.808
#> 15:        7--8   hetero        0       2.571 -1.83632813         1.000
#> 16:        3--6   hetero       32      55.818 -0.78388202         1.000
#> 17:        4--6   hetero       20      39.911 -0.96209947         1.000
#> 18:        1--2   hetero       83     167.024 -1.00020608         1.000
#> 19:        6--8   hetero        0       3.350 -2.12101540         1.000
#> 20:        3--8   hetero        1       5.798 -1.76511036         0.997
#> 21:        5--8   hetero        0       3.711 -2.23603333         1.000
#> 22:        4--8   hetero        0       4.141 -2.36204901         1.000
#> 23:        1--7   hetero       20      61.010 -1.56211156         1.000
#> 24:        1--4   hetero       29      98.163 -1.72483942         1.000
#> 25:        6--7   hetero        3      24.975 -2.69905184         1.000
#> 26:        3--4   hetero        6      65.457 -3.24699434         1.000
#> 27:        2--3   hetero        9     111.922 -3.49725468         1.000
#> 28:        1--5   hetero        3      92.113 -4.54091070         1.000
#> 29:        2--7   hetero        1      49.632 -4.66197757         1.000
#> 30:        5--7   hetero        0      27.183 -4.81675329         1.000
#> 31:        4--7   hetero        0      28.873 -4.90077022         1.000
#> 32:        2--6   hetero        1      67.441 -5.09678893         1.000
#> 33:        5--6   hetero        0      37.355 -5.26134276         1.000
#> 34:        3--5   hetero        0      61.577 -5.96756059         1.000
#> 35:        2--5   hetero        0      75.017 -6.24825019         1.000
#> 36:        2--4   hetero        0      79.649 -6.33358474         1.000
#>     unified_int type_int original simulations     enrichm p_higher_orig
#>     p_lower_orig p.adj_higher p.adj_lower     PI_value int_ranking
#>            <num>        <num>       <num>        <num>       <int>
#>  1:        1.000        0.000       1.000   8.12690410           1
#>  2:        1.000        0.000       1.000   7.70342102           2
#>  3:        1.000        0.000       1.000   7.47773275           3
#>  4:        1.000        0.000       1.000   6.49914819           4
#>  5:        1.000        0.000       1.000   6.34160277           5
#>  6:        1.000        0.000       1.000   5.32787918           6
#>  7:        1.000        0.000       1.000   4.88012665           7
#>  8:        1.000        0.000       1.000   4.37415328           8
#>  9:        1.000        0.000       1.000   4.02479423           9
#> 10:        1.000        0.000       1.000   3.79936428          10
#> 11:        1.000        0.000       1.000   3.34264703          11
#> 12:        0.999        0.001       0.999   0.76228804          12
#> 13:        0.359        0.692       0.359  -0.02627511          13
#> 14:        0.336        0.808       0.336  -0.16315492          14
#> 15:        0.062        1.000       0.062  -2.20480473          15
#> 16:        0.000        1.000       0.000  -2.35164607          16
#> 17:        0.000        1.000       0.000  -2.88629842          17
#> 18:        0.000        1.000       0.000  -3.00061825          18
#> 19:        0.031        1.000       0.031  -3.17059992          19
#> 20:        0.013        0.997       0.013  -3.27228861          20
#> 21:        0.020        1.000       0.020  -3.75157358          21
#> 22:        0.011        1.000       0.011  -4.53706804          22
#> 23:        0.000        1.000       0.000  -4.68633468          23
#> 24:        0.000        1.000       0.000  -5.17451825          24
#> 25:        0.000        1.000       0.000  -8.09715553          25
#> 26:        0.000        1.000       0.000  -9.74098302          26
#> 27:        0.000        1.000       0.000 -10.49176404          27
#> 28:        0.000        1.000       0.000 -13.62273210          28
#> 29:        0.000        1.000       0.000 -13.98593270          29
#> 30:        0.000        1.000       0.000 -14.45025986          30
#> 31:        0.000        1.000       0.000 -14.70231067          31
#> 32:        0.000        1.000       0.000 -15.29036680          32
#> 33:        0.000        1.000       0.000 -15.78402827          33
#> 34:        0.000        1.000       0.000 -17.90268177          34
#> 35:        0.000        1.000       0.000 -18.74475056          35
#> 36:        0.000        1.000       0.000 -19.00075422          36
#>     p_lower_orig p.adj_higher p.adj_lower     PI_value int_ranking
#>