cellProximityEnrichmentSpots

Compute cell-cell interaction enrichment for spots (observed vs expected)

cellProximityEnrichmentSpots(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "spatial_network",
  cluster_column = "cell_ID",
  cells_in_spot = 1,
  number_of_simulations = 100,
  adjust_method = c("none", "fdr", "bonferroni", "BH", "holm", "hochberg", "hommel",
    "BY"),
  set_seed = TRUE,
  seed_number = 1234,
  verbose = FALSE
)

Arguments

gobject

giotto object

spat_unit

spatial unit (e.g. 'cell')

feat_type

feature type (e.g. 'rna')

spatial_network_name

name of spatial network to use

cluster_column

name of column to use for clusters

cells_in_spot

cell number in each spot

number_of_simulations

number of simulations to create expected observations

adjust_method

method to adjust p.values (e.g. "none", "fdr", "bonferroni","BH","holm", "hochberg", "hommel","BY")

set_seed

use of seed. Default = TRUE

seed_number

seed number to use. Default = 1234

verbose

be verbose

Value

List of cell Proximity scores (CPscores) in data.table format. The first data.table (raw_sim_table) shows the raw observations of both the original and simulated networks. The second data.table (enrichm_res) shows the enrichment results.

Details

Spatial proximity enrichment or depletion between pairs of cell types is calculated by calculating the observed over the expected frequency of cell-cell proximity interactions. The expected frequency is the average frequency calculated from a number of spatial network simulations. Each individual simulation is obtained by reshuffling the cell type labels of each node (spot) in the spatial network.

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used
x <- findMarkers_one_vs_all(g,
cluster_column = "leiden_clus", min_feats = 20)
#> using 'Scran' to detect marker feats. If used in published 
#>       research, please cite: Lun ATL, McCarthy DJ, Marioni JC (2016).
#>       'A step-by-step workflow for low-level analysis of single-cell RNA-seq 
#>       data with Bioconductor.'
#>       F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2. 
#> start with cluster  1start with cluster  2start with cluster  3start with cluster  4start with cluster  5start with cluster  6start with cluster  7start with cluster  8
sign_gene <- x$feats

sign_matrix <- matrix(rnorm(length(sign_gene)*8, mean  = 10),
nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- paste0("celltype_",unique(x$cluster))

g <- runDWLSDeconv(gobject = g, sign_matrix = sign_matrix)

cellProximityEnrichmentSpots(gobject = g)
#> $raw_sim_table
#> Index: <unified_int>
#>                  unified_int type_int        V1  external   internal
#>                       <char>   <char>     <num>     <num>      <num>
#>    1: celltype_1--celltype_1     homo  31.73189  31.73189 0.00000000
#>    2: celltype_1--celltype_2   hetero  45.71553  44.78616 0.92936377
#>    3: celltype_1--celltype_3   hetero  47.87709  47.84967 0.02741609
#>    4: celltype_1--celltype_4   hetero  41.53840  41.37635 0.16204178
#>    5: celltype_1--celltype_5   hetero  47.77458  45.68953 2.08505107
#>   ---                                                               
#> 3632: celltype_6--celltype_7   hetero 113.58794 113.48799 0.09995237
#> 3633: celltype_6--celltype_8   hetero  40.07812  40.07812 0.00000000
#> 3634: celltype_7--celltype_7     homo 230.34240 230.34240 0.00000000
#> 3635: celltype_7--celltype_8   hetero 120.49663 120.49663 0.00000000
#> 3636: celltype_8--celltype_8     homo  44.98556  44.98556 0.00000000
#>              orig    round
#>            <char>   <char>
#>    1:    original original
#>    2:    original original
#>    3:    original original
#>    4:    original original
#>    5:    original original
#>   ---                     
#> 3632: simulations   sim100
#> 3633: simulations   sim100
#> 3634: simulations   sim100
#> 3635: simulations   sim100
#> 3636: simulations   sim100
#> 
#> $enrichm_res
#>                unified_int type_int   original simulations     enrichm
#>                     <fctr>   <char>      <num>       <num>       <num>
#>  1: celltype_1--celltype_3   hetero  47.877090    19.43156  1.25835918
#>  2: celltype_3--celltype_8   hetero  57.476190    32.85346  0.78854606
#>  3: celltype_1--celltype_1     homo  31.731890    19.34873  0.68575811
#>  4: celltype_3--celltype_3     homo  43.638299    30.35388  0.50963822
#>  5: celltype_1--celltype_2   hetero  45.715527    33.09691  0.45426130
#>  6: celltype_2--celltype_3   hetero  61.174659    44.64674  0.44581473
#>  7: celltype_2--celltype_2     homo  87.546785    66.06409  0.40089948
#>  8: celltype_4--celltype_4     homo 265.159201   207.13477  0.35477143
#>  9: celltype_1--celltype_8   hetero  31.881141    24.10807  0.38910935
#> 10: celltype_2--celltype_5   hetero 157.421489   127.00996  0.30751198
#> 11: celltype_6--celltype_7   hetero 139.784075   113.04652  0.30386176
#> 12: celltype_7--celltype_7     homo 294.304722   239.51775  0.29606107
#> 13: celltype_4--celltype_6   hetero 132.349030   109.33230  0.27335218
#> 14: celltype_5--celltype_5     homo 199.542192   165.96131  0.26439200
#> 15: celltype_8--celltype_8     homo  48.253830    43.69978  0.13996830
#> 16: celltype_7--celltype_8   hetero 129.068997   120.13159  0.10270200
#> 17: celltype_5--celltype_6   hetero  98.315676    92.18995  0.09184712
#> 18: celltype_4--celltype_5   hetero 275.057166   261.89425  0.07048448
#> 19: celltype_6--celltype_6     homo  38.843156    39.42028 -0.02074730
#> 20: celltype_2--celltype_8   hetero  51.392305    54.51004 -0.08339370
#> 21: celltype_4--celltype_7   hetero 314.542337   327.25192 -0.05696985
#> 22: celltype_6--celltype_8   hetero  33.638525    40.41575 -0.25780199
#> 23: celltype_3--celltype_7   hetero  82.388195    96.11624 -0.21986947
#> 24: celltype_2--celltype_6   hetero  43.856684    52.59783 -0.25685159
#> 25: celltype_1--celltype_5   hetero  47.774582    58.73320 -0.29240353
#> 26: celltype_4--celltype_8   hetero  96.234848   113.83593 -0.24002871
#> 27: celltype_5--celltype_8   hetero  80.244784    98.01639 -0.28539211
#> 28: celltype_5--celltype_7   hetero 233.207299   280.62315 -0.26597987
#> 29: celltype_2--celltype_7   hetero 130.258754   162.82694 -0.31975899
#> 30: celltype_1--celltype_7   hetero  52.467842    71.57382 -0.44077785
#> 31: celltype_2--celltype_4   hetero 102.669677   149.12576 -0.53417761
#> 32: celltype_3--celltype_5   hetero  51.804085    77.21398 -0.56677699
#> 33: celltype_3--celltype_4   hetero  59.082852    91.03279 -0.61519463
#> 34: celltype_1--celltype_4   hetero  41.538397    64.76748 -0.62860882
#> 35: celltype_3--celltype_6   hetero  13.461038    31.45666 -1.16634357
#> 36: celltype_1--celltype_6   hetero   8.779884    23.60069 -1.33080978
#>                unified_int type_int   original simulations     enrichm
#>     p_higher_orig p_lower_orig p.adj_higher p.adj_lower     PI_value
#>             <num>        <num>        <num>       <num>        <num>
#>  1:          0.00         1.00         0.00        1.00  2.516718370
#>  2:          0.00         1.00         0.00        1.00  1.577092114
#>  3:          0.00         1.00         0.00        1.00  1.371516223
#>  4:          0.00         1.00         0.00        1.00  1.019276434
#>  5:          0.00         1.00         0.00        1.00  0.908522596
#>  6:          0.00         1.00         0.00        1.00  0.891629452
#>  7:          0.00         1.00         0.00        1.00  0.801798952
#>  8:          0.00         1.00         0.00        1.00  0.709542853
#>  9:          0.01         0.99         0.01        0.99  0.661085122
#> 10:          0.00         1.00         0.00        1.00  0.615023951
#> 11:          0.00         1.00         0.00        1.00  0.607723516
#> 12:          0.00         1.00         0.00        1.00  0.592122142
#> 13:          0.00         1.00         0.00        1.00  0.546704357
#> 14:          0.00         1.00         0.00        1.00  0.528783998
#> 15:          0.10         0.90         0.10        0.90  0.134174635
#> 16:          0.13         0.87         0.13        0.87  0.087694356
#> 17:          0.24         0.76         0.24        0.76  0.055297478
#> 18:          0.16         0.84         0.16        0.84  0.054241406
#> 19:          0.54         0.46         0.54        0.46 -0.006803083
#> 20:          0.68         0.32         0.68        0.32 -0.040152902
#> 21:          0.84         0.16         0.84        0.16 -0.043841211
#> 22:          0.91         0.09         0.91        0.09 -0.257801988
#> 23:          0.97         0.03         0.97        0.03 -0.307364330
#> 24:          0.96         0.04         0.96        0.04 -0.334171627
#> 25:          0.98         0.02         0.98        0.02 -0.445295123
#> 26:          1.00         0.00         1.00        0.00 -0.480057413
#> 27:          0.99         0.01         0.99        0.01 -0.484872641
#> 28:          1.00         0.00         1.00        0.00 -0.531959738
#> 29:          1.00         0.00         1.00        0.00 -0.639517980
#> 30:          1.00         0.00         1.00        0.00 -0.881555697
#> 31:          1.00         0.00         1.00        0.00 -1.068355218
#> 32:          1.00         0.00         1.00        0.00 -1.133553988
#> 33:          1.00         0.00         1.00        0.00 -1.230389258
#> 34:          1.00         0.00         1.00        0.00 -1.257217638
#> 35:          1.00         0.00         1.00        0.00 -2.332687133
#> 36:          1.00         0.00         1.00        0.00 -2.661619560
#>     p_higher_orig p_lower_orig p.adj_higher p.adj_lower     PI_value
#>     int_ranking
#>           <int>
#>  1:           1
#>  2:           2
#>  3:           3
#>  4:           4
#>  5:           5
#>  6:           6
#>  7:           7
#>  8:           8
#>  9:           9
#> 10:          10
#> 11:          11
#> 12:          12
#> 13:          13
#> 14:          14
#> 15:          15
#> 16:          16
#> 17:          17
#> 18:          18
#> 19:          19
#> 20:          20
#> 21:          21
#> 22:          22
#> 23:          23
#> 24:          24
#> 25:          25
#> 26:          26
#> 27:          27
#> 28:          28
#> 29:          29
#> 30:          30
#> 31:          31
#> 32:          32
#> 33:          33
#> 34:          34
#> 35:          35
#> 36:          36
#>     int_ranking
#>