exprCellCellcom

Cell-Cell communication scores based on expression only

exprCellCellcom(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  cluster_column = "cell_types",
  random_iter = 1000,
  feat_set_1,
  feat_set_2,
  log2FC_addendum = 0.1,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("feats", "cells"),
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE
)

Arguments

gobject

giotto object to use

feat_type

feature type

spat_unit

spatial unit

cluster_column

cluster column with cell type information

random_iter

number of iterations

feat_set_1

first specific feature set from feature pairs

feat_set_2

second specific feature set from feature pairs

log2FC_addendum

addendum to add when calculating log2FC

detailed

provide more detailed information (random variance and z-score)

adjust_method

which method to adjust p-values

adjust_target

adjust multiple hypotheses at the cell or feature level

set_seed

set seed for random simulations (default = TRUE)

seed_number

seed number

verbose

verbose

Value

Cell-Cell communication scores for feature pairs based on expression only

Details

Statistical framework to identify if pairs of features (such as ligand-receptor combinations) are expressed at higher levels than expected based on a reshuffled null distribution of feature expression values, without considering the spatial position of cells. More details will follow soon.

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path or giotto
#>  environment was specified, will check if a
#>  default python path is available
#> 
#> A default python path was found: /usr/bin/python3 and will be used
#>  If this is not the correct python path, either
#> 
#> 1. use installGiottoEnvironment() to install
#>  a local miniconda python environment along with required modules
#> 
#> 2. provide an existing python path to
#>  python_path to use your own python path which has all modules
#>  installed
#> Set options("giotto.use_conda" = FALSE) if
#>  python functionalities are not needed
#> Warning: This giotto object was created in a newer version of
#>                 GiottoClass (v0.3.2)

exprCellCellcom(g, cluster_column = "leiden_clus",
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik")
#> simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  11simulation  12simulation  13simulation  14simulation  15simulation  16simulation  17simulation  18simulation  19simulation  20simulation  21simulation  22simulation  23simulation  24simulation  25simulation  26simulation  27simulation  28simulation  29simulation  30simulation  31simulation  32simulation  33simulation  34simulation  35simulation  36simulation  37simulation  38simulation  39simulation  40simulation  41simulation  42simulation  43simulation  44simulation  45simulation  46simulation  47simulation  48simulation  49simulation  50simulation  51simulation  52simulation  53simulation  54simulation  55simulation  56simulation  57simulation  58simulation  59simulation  60simulation  61simulation  62simulation  63simulation  64simulation  65simulation  66simulation  67simulation  68simulation  69simulation  70simulation  71simulation  72simulation  73simulation  74simulation  75simulation  76simulation  77simulation  78simulation  79simulation  80simulation  81simulation  82simulation  83simulation  84simulation  85simulation  86simulation  87simulation  88simulation  89simulation  90simulation  91simulation  92simulation  93simulation  94simulation  95simulation  96simulation  97simulation  98simulation  99simulation  100simulation  101simulation  102simulation  103simulation  104simulation  105simulation  106simulation  107simulation  108simulation  109simulation  110simulation  111simulation  112simulation  113simulation  114simulation  115simulation  116simulation  117simulation  118simulation  119simulation  120simulation  121simulation  122simulation  123simulation  124simulation  125simulation  126simulation  127simulation  128simulation  129simulation  130simulation  131simulation  132simulation  133simulation  134simulation  135simulation  136simulation  137simulation  138simulation  139simulation  140simulation  141simulation  142simulation  143simulation  144simulation  145simulation  146simulation  147simulation  148simulation  149simulation  150simulation  151simulation  152simulation  153simulation  154simulation  155simulation  156simulation  157simulation  158simulation  159simulation  160simulation  161simulation  162simulation  163simulation  164simulation  165simulation  166simulation  167simulation  168simulation  169simulation  170simulation  171simulation  172simulation  173simulation  174simulation  175simulation  176simulation  177simulation  178simulation  179simulation  180simulation  181simulation  182simulation  183simulation  184simulation  185simulation  186simulation  187simulation  188simulation  189simulation  190simulation  191simulation  192simulation  193simulation  194simulation  195simulation  196simulation  197simulation  198simulation  199simulation  200simulation  201simulation  202simulation  203simulation  204simulation  205simulation  206simulation  207simulation  208simulation  209simulation  210simulation  211simulation  212simulation  213simulation  214simulation  215simulation  216simulation  217simulation  218simulation  219simulation  220simulation  221simulation  222simulation  223simulation  224simulation  225simulation  226simulation  227simulation  228simulation  229simulation  230simulation  231simulation  232simulation  233simulation  234simulation  235simulation  236simulation  237simulation  238simulation  239simulation  240simulation  241simulation  242simulation  243simulation  244simulation  245simulation  246simulation  247simulation  248simulation  249simulation  250simulation  251simulation  252simulation  253simulation  254simulation  255simulation  256simulation  257simulation  258simulation  259simulation  260simulation  261simulation  262simulation  263simulation  264simulation  265simulation  266simulation  267simulation  268simulation  269simulation  270simulation  271simulation  272simulation  273simulation  274simulation  275simulation  276simulation  277simulation  278simulation  279simulation  280simulation  281simulation  282simulation  283simulation  284simulation  285simulation  286simulation  287simulation  288simulation  289simulation  290simulation  291simulation  292simulation  293simulation  294simulation  295simulation  296simulation  297simulation  298simulation  299simulation  300simulation  301simulation  302simulation  303simulation  304simulation  305simulation  306simulation  307simulation  308simulation  309simulation  310simulation  311simulation  312simulation  313simulation  314simulation  315simulation  316simulation  317simulation  318simulation  319simulation  320simulation  321simulation  322simulation  323simulation  324simulation  325simulation  326simulation  327simulation  328simulation  329simulation  330simulation  331simulation  332simulation  333simulation  334simulation  335simulation  336simulation  337simulation  338simulation  339simulation  340simulation  341simulation  342simulation  343simulation  344simulation  345simulation  346simulation  347simulation  348simulation  349simulation  350simulation  351simulation  352simulation  353simulation  354simulation  355simulation  356simulation  357simulation  358simulation  359simulation  360simulation  361simulation  362simulation  363simulation  364simulation  365simulation  366simulation  367simulation  368simulation  369simulation  370simulation  371simulation  372simulation  373simulation  374simulation  375simulation  376simulation  377simulation  378simulation  379simulation  380simulation  381simulation  382simulation  383simulation  384simulation  385simulation  386simulation  387simulation  388simulation  389simulation  390simulation  391simulation  392simulation  393simulation  394simulation  395simulation  396simulation  397simulation  398simulation  399simulation  400simulation  401simulation  402simulation  403simulation  404simulation  405simulation  406simulation  407simulation  408simulation  409simulation  410simulation  411simulation  412simulation  413simulation  414simulation  415simulation  416simulation  417simulation  418simulation  419simulation  420simulation  421simulation  422simulation  423simulation  424simulation  425simulation  426simulation  427simulation  428simulation  429simulation  430simulation  431simulation  432simulation  433simulation  434simulation  435simulation  436simulation  437simulation  438simulation  439simulation  440simulation  441simulation  442simulation  443simulation  444simulation  445simulation  446simulation  447simulation  448simulation  449simulation  450simulation  451simulation  452simulation  453simulation  454simulation  455simulation  456simulation  457simulation  458simulation  459simulation  460simulation  461simulation  462simulation  463simulation  464simulation  465simulation  466simulation  467simulation  468simulation  469simulation  470simulation  471simulation  472simulation  473simulation  474simulation  475simulation  476simulation  477simulation  478simulation  479simulation  480simulation  481simulation  482simulation  483simulation  484simulation  485simulation  486simulation  487simulation  488simulation  489simulation  490simulation  491simulation  492simulation  493simulation  494simulation  495simulation  496simulation  497simulation  498simulation  499simulation  500simulation  501simulation  502simulation  503simulation  504simulation  505simulation  506simulation  507simulation  508simulation  509simulation  510simulation  511simulation  512simulation  513simulation  514simulation  515simulation  516simulation  517simulation  518simulation  519simulation  520simulation  521simulation  522simulation  523simulation  524simulation  525simulation  526simulation  527simulation  528simulation  529simulation  530simulation  531simulation  532simulation  533simulation  534simulation  535simulation  536simulation  537simulation  538simulation  539simulation  540simulation  541simulation  542simulation  543simulation  544simulation  545simulation  546simulation  547simulation  548simulation  549simulation  550simulation  551simulation  552simulation  553simulation  554simulation  555simulation  556simulation  557simulation  558simulation  559simulation  560simulation  561simulation  562simulation  563simulation  564simulation  565simulation  566simulation  567simulation  568simulation  569simulation  570simulation  571simulation  572simulation  573simulation  574simulation  575simulation  576simulation  577simulation  578simulation  579simulation  580simulation  581simulation  582simulation  583simulation  584simulation  585simulation  586simulation  587simulation  588simulation  589simulation  590simulation  591simulation  592simulation  593simulation  594simulation  595simulation  596simulation  597simulation  598simulation  599simulation  600simulation  601simulation  602simulation  603simulation  604simulation  605simulation  606simulation  607simulation  608simulation  609simulation  610simulation  611simulation  612simulation  613simulation  614simulation  615simulation  616simulation  617simulation  618simulation  619simulation  620simulation  621simulation  622simulation  623simulation  624simulation  625simulation  626simulation  627simulation  628simulation  629simulation  630simulation  631simulation  632simulation  633simulation  634simulation  635simulation  636simulation  637simulation  638simulation  639simulation  640simulation  641simulation  642simulation  643simulation  644simulation  645simulation  646simulation  647simulation  648simulation  649simulation  650simulation  651simulation  652simulation  653simulation  654simulation  655simulation  656simulation  657simulation  658simulation  659simulation  660simulation  661simulation  662simulation  663simulation  664simulation  665simulation  666simulation  667simulation  668simulation  669simulation  670simulation  671simulation  672simulation  673simulation  674simulation  675simulation  676simulation  677simulation  678simulation  679simulation  680simulation  681simulation  682simulation  683simulation  684simulation  685simulation  686simulation  687simulation  688simulation  689simulation  690simulation  691simulation  692simulation  693simulation  694simulation  695simulation  696simulation  697simulation  698simulation  699simulation  700simulation  701simulation  702simulation  703simulation  704simulation  705simulation  706simulation  707simulation  708simulation  709simulation  710simulation  711simulation  712simulation  713simulation  714simulation  715simulation  716simulation  717simulation  718simulation  719simulation  720simulation  721simulation  722simulation  723simulation  724simulation  725simulation  726simulation  727simulation  728simulation  729simulation  730simulation  731simulation  732simulation  733simulation  734simulation  735simulation  736simulation  737simulation  738simulation  739simulation  740simulation  741simulation  742simulation  743simulation  744simulation  745simulation  746simulation  747simulation  748simulation  749simulation  750simulation  751simulation  752simulation  753simulation  754simulation  755simulation  756simulation  757simulation  758simulation  759simulation  760simulation  761simulation  762simulation  763simulation  764simulation  765simulation  766simulation  767simulation  768simulation  769simulation  770simulation  771simulation  772simulation  773simulation  774simulation  775simulation  776simulation  777simulation  778simulation  779simulation  780simulation  781simulation  782simulation  783simulation  784simulation  785simulation  786simulation  787simulation  788simulation  789simulation  790simulation  791simulation  792simulation  793simulation  794simulation  795simulation  796simulation  797simulation  798simulation  799simulation  800simulation  801simulation  802simulation  803simulation  804simulation  805simulation  806simulation  807simulation  808simulation  809simulation  810simulation  811simulation  812simulation  813simulation  814simulation  815simulation  816simulation  817simulation  818simulation  819simulation  820simulation  821simulation  822simulation  823simulation  824simulation  825simulation  826simulation  827simulation  828simulation  829simulation  830simulation  831simulation  832simulation  833simulation  834simulation  835simulation  836simulation  837simulation  838simulation  839simulation  840simulation  841simulation  842simulation  843simulation  844simulation  845simulation  846simulation  847simulation  848simulation  849simulation  850simulation  851simulation  852simulation  853simulation  854simulation  855simulation  856simulation  857simulation  858simulation  859simulation  860simulation  861simulation  862simulation  863simulation  864simulation  865simulation  866simulation  867simulation  868simulation  869simulation  870simulation  871simulation  872simulation  873simulation  874simulation  875simulation  876simulation  877simulation  878simulation  879simulation  880simulation  881simulation  882simulation  883simulation  884simulation  885simulation  886simulation  887simulation  888simulation  889simulation  890simulation  891simulation  892simulation  893simulation  894simulation  895simulation  896simulation  897simulation  898simulation  899simulation  900simulation  901simulation  902simulation  903simulation  904simulation  905simulation  906simulation  907simulation  908simulation  909simulation  910simulation  911simulation  912simulation  913simulation  914simulation  915simulation  916simulation  917simulation  918simulation  919simulation  920simulation  921simulation  922simulation  923simulation  924simulation  925simulation  926simulation  927simulation  928simulation  929simulation  930simulation  931simulation  932simulation  933simulation  934simulation  935simulation  936simulation  937simulation  938simulation  939simulation  940simulation  941simulation  942simulation  943simulation  944simulation  945simulation  946simulation  947simulation  948simulation  949simulation  950simulation  951simulation  952simulation  953simulation  954simulation  955simulation  956simulation  957simulation  958simulation  959simulation  960simulation  961simulation  962simulation  963simulation  964simulation  965simulation  966simulation  967simulation  968simulation  969simulation  970simulation  971simulation  972simulation  973simulation  974simulation  975simulation  976simulation  977simulation  978simulation  979simulation  980simulation  981simulation  982simulation  983simulation  984simulation  985simulation  986simulation  987simulation  988simulation  989simulation  990simulation  991simulation  992simulation  993simulation  994simulation  995simulation  996simulation  997simulation  998simulation  999simulation  1000
#> Key: <LR_comb>
#>                   LR_comb lig_cell_type  lig_expr  ligand rec_cell_type
#>                    <char>        <fctr>     <num>  <char>        <fctr>
#>  1: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             4
#>  2: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             5
#>  3: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             1
#>  4: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             3
#>  5: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             4
#>  6: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             5
#>  7: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             1
#>  8: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             3
#>  9: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             4
#> 10: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             2
#> 11: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             5
#> 12: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             1
#> 13: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             4
#> 14: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             4
#> 15: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             3
#> 16: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             4
#> 17: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             5
#> 18: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             1
#> 19: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             2
#> 20: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             6
#> 21: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             5
#> 22: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             1
#> 23: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             3
#> 24: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             5
#> 25: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             1
#> 26: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             3
#> 27: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             3
#> 28: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             2
#> 29: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             6
#> 30: Gm19935-9630013A20Rik             6 0.4293181 Gm19935             7
#> 31: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             2
#> 32: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             2
#> 33: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             2
#> 34: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             6
#> 35: Gm19935-9630013A20Rik             2 0.3574502 Gm19935             7
#> 36: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             6
#> 37: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             6
#> 38: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             4
#> 39: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             6
#> 40: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             5
#> 41: Gm19935-9630013A20Rik             5 0.2922187 Gm19935             7
#> 42: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             1
#> 43: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             3
#> 44: Gm19935-9630013A20Rik             4 0.2495580 Gm19935             7
#> 45: Gm19935-9630013A20Rik             1 0.2326362 Gm19935             7
#> 46: Gm19935-9630013A20Rik             3 0.2197897 Gm19935             7
#> 47: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             2
#> 48: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             6
#> 49: Gm19935-9630013A20Rik             7 0.0000000 Gm19935             7
#>                   LR_comb lig_cell_type  lig_expr  ligand rec_cell_type
#>       rec_expr      receptor    LR_expr lig_nr rec_nr rand_expr      av_diff
#>          <num>        <char>      <num>  <int>  <int>     <num>        <num>
#>  1: 0.33023852 9630013A20Rik 0.75955660     10    108 0.5247843  0.234772341
#>  2: 0.29575033 9630013A20Rik 0.72506842     10     45 0.5262755  0.198792953
#>  3: 0.28874834 9630013A20Rik 0.71806642     10    122 0.5237300  0.194336461
#>  4: 0.26833691 9630013A20Rik 0.69765500     10    162 0.5254247  0.172230266
#>  5: 0.33023852 9630013A20Rik 0.68768868     84    108 0.5266116  0.161077071
#>  6: 0.29575033 9630013A20Rik 0.65320049     84     45 0.5281028  0.125097683
#>  7: 0.28874834 9630013A20Rik 0.64619850     84    122 0.5255573  0.120641191
#>  8: 0.26833691 9630013A20Rik 0.62578707     84    162 0.5272521  0.098534996
#>  9: 0.33023852 9630013A20Rik 0.62245718     45    108 0.5329486  0.089508532
#> 10: 0.17597264 9630013A20Rik 0.60529072     10     84 0.5246925  0.080598186
#> 11: 0.29575033 9630013A20Rik 0.58796899     45     45 0.5344398  0.053529145
#> 12: 0.28874834 9630013A20Rik 0.58096700     45    122 0.5318943  0.049072652
#> 13: 0.33023852 9630013A20Rik 0.57979653    108    108 0.5281585  0.051637977
#> 14: 0.33023852 9630013A20Rik 0.56287473    122    108 0.5279882  0.034886510
#> 15: 0.26833691 9630013A20Rik 0.56055557     45    162 0.5335891  0.026966457
#> 16: 0.33023852 9630013A20Rik 0.55002826    162    108 0.5269931  0.023035131
#> 17: 0.29575033 9630013A20Rik 0.54530835    108     45 0.5296498  0.015658590
#> 18: 0.28874834 9630013A20Rik 0.53830635    108    122 0.5271043  0.011202097
#> 19: 0.17597264 9630013A20Rik 0.53342280     84     84 0.5265199  0.006902916
#> 20: 0.09952757 9630013A20Rik 0.52884565     10     10 0.5209773  0.007868373
#> 21: 0.29575033 9630013A20Rik 0.52838655    122     45 0.5294794 -0.001092878
#> 22: 0.28874834 9630013A20Rik 0.52138455    122    122 0.5269339 -0.005549370
#> 23: 0.26833691 9630013A20Rik 0.51789492    108    162 0.5287990 -0.010904098
#> 24: 0.29575033 9630013A20Rik 0.51554007    162     45 0.5284843 -0.012944257
#> 25: 0.28874834 9630013A20Rik 0.50853808    162    122 0.5259388 -0.017400749
#> 26: 0.26833691 9630013A20Rik 0.50097313    122    162 0.5286287 -0.027655565
#> 27: 0.26833691 9630013A20Rik 0.48812665    162    162 0.5276336 -0.039506944
#> 28: 0.17597264 9630013A20Rik 0.46819130     45     84 0.5328569 -0.064665622
#> 29: 0.09952757 9630013A20Rik 0.45697772     84     10 0.5228046 -0.065826897
#> 30: 0.00000000 9630013A20Rik 0.42931808     10     93 0.5204358 -0.091117732
#> 31: 0.17597264 9630013A20Rik 0.42553065    108     84 0.5280668 -0.102536177
#> 32: 0.17597264 9630013A20Rik 0.40860885    122     84 0.5278965 -0.119287645
#> 33: 0.17597264 9630013A20Rik 0.39576238    162     84 0.5269014 -0.131139024
#> 34: 0.09952757 9630013A20Rik 0.39174622     45     10 0.5291417 -0.137395436
#> 35: 0.00000000 9630013A20Rik 0.35745016     84     93 0.5222632 -0.164813002
#> 36: 0.09952757 9630013A20Rik 0.34908558    108     10 0.5243516 -0.175265991
#> 37: 0.09952757 9630013A20Rik 0.33216378    122     10 0.5241812 -0.192017458
#> 38: 0.33023852 9630013A20Rik 0.33023852     93    108 0.5372749 -0.207036394
#> 39: 0.09952757 9630013A20Rik 0.31931730    162     10 0.5231861 -0.203868837
#> 40: 0.29575033 9630013A20Rik 0.29575033     93     45 0.5387661 -0.243015781
#> 41: 0.00000000 9630013A20Rik 0.29221866     45     93 0.5286002 -0.236381541
#> 42: 0.28874834 9630013A20Rik 0.28874834     93    122 0.5362206 -0.247472274
#> 43: 0.26833691 9630013A20Rik 0.26833691     93    162 0.5379154 -0.269578469
#> 44: 0.00000000 9630013A20Rik 0.24955801    108     93 0.5238101 -0.274252096
#> 45: 0.00000000 9630013A20Rik 0.23263621    122     93 0.5236398 -0.291003563
#> 46: 0.00000000 9630013A20Rik 0.21978974    162     93 0.5226447 -0.302854942
#> 47: 0.17597264 9630013A20Rik 0.17597264     93     84 0.5371832 -0.361210548
#> 48: 0.09952757 9630013A20Rik 0.09952757     93     10 0.5334679 -0.433940362
#> 49: 0.00000000 9630013A20Rik 0.00000000     93     93 0.5329265 -0.532926467
#>       rec_expr      receptor    LR_expr lig_nr rec_nr rand_expr      av_diff
#>           log2fc pvalue LR_cell_comb p.adj            PI
#>            <num>  <num>       <char> <num>         <num>
#>  1:  0.460234536  0.150         6--4 0.150  3.791913e-01
#>  2:  0.397716399  0.236         6--5 0.236  2.494032e-01
#>  3:  0.391296425  0.192         6--1 0.192  2.804417e-01
#>  4:  0.350928616  0.270         6--3 0.270  1.995507e-01
#>  5:  0.330054046  0.162         2--4 0.162  2.609028e-01
#>  6:  0.262033221  0.304         2--5 0.304  1.355043e-01
#>  7:  0.254417400  0.184         2--1 0.184  1.870431e-01
#>  8:  0.210501018  0.364         2--3 0.364  9.238861e-02
#>  9:  0.190823627  0.468         5--4 0.468  6.292488e-02
#> 10:  0.175071769  0.606         6--2 0.606  3.808290e-02
#> 11:  0.116860156  0.696         5--5 0.696  1.839271e-02
#> 12:  0.107901525  0.644         5--1 0.644  2.062151e-02
#> 13:  0.113974244  0.676         4--4 0.676  1.938170e-02
#> 14:  0.077998759  0.698         1--4 0.698  1.217908e-02
#> 15:  0.060132389  0.810         5--3 0.810  5.503014e-03
#> 16:  0.052052807  0.814         3--4 0.814  4.652251e-03
#> 17:  0.035439134  0.888         4--5 0.888  1.828200e-03
#> 18:  0.025543697  0.904         4--1 0.904  1.119620e-03
#> 19:  0.015808501  0.926         2--2 0.926  5.278303e-04
#> 20:  0.018165474  0.942         6--6 0.942  4.713776e-04
#> 21: -0.002506928  0.960         1--5 0.960 -4.444474e-05
#> 22: -0.012827017  0.948         1--1 0.948 -2.974799e-04
#> 23: -0.025237452  0.950         4--3 0.950 -5.621994e-04
#> 24: -0.030023992  0.944         3--5 0.944 -7.514406e-04
#> 25: -0.040674131  0.888         3--1 0.888 -2.098258e-03
#> 26: -0.064907641  0.792         1--3 0.792 -6.573510e-03
#> 27: -0.093795710  0.784         3--3 0.784 -9.912700e-03
#> 28: -0.155502627  0.658         5--2 0.658 -2.826635e-02
#> 29: -0.161160019  0.700         2--6 0.700 -2.496400e-02
#> 30: -0.229147029  0.894         6--7 0.894 -1.115086e-02
#> 31: -0.257143164  0.396         4--2 0.396 -1.034499e-01
#> 32: -0.303970191  0.188         1--2 0.188 -2.206344e-01
#> 33: -0.338589761  0.214         3--2 0.214 -2.267150e-01
#> 34: -0.355470921  0.380         5--6 0.380 -1.493747e-01
#> 35: -0.443910274  0.758         2--7 0.758 -5.341608e-02
#> 36: -0.475368241  0.286         4--6 0.286 -2.584263e-01
#> 37: -0.530386832  0.166         1--6 0.166 -4.136444e-01
#> 38: -0.566779176  0.642         7--4 0.642 -1.090851e-01
#> 39: -0.571620783  0.148         3--6 0.148 -4.742956e-01
#> 40: -0.690697215  0.532         7--5 0.532 -1.893121e-01
#> 41: -0.680484559  0.500         5--7 0.500 -2.048463e-01
#> 42: -0.710690600  0.522         7--1 0.522 -2.006489e-01
#> 43: -0.792339080  0.442         7--3 0.442 -2.809458e-01
#> 44: -0.835575030  0.356         4--7 0.356 -3.747972e-01
#> 45: -0.906767711  0.214         1--7 0.214 -6.071592e-01
#> 46: -0.961285460  0.216         3--7 0.216 -6.397799e-01
#> 47: -1.207182952  0.118         7--2 0.118 -1.120408e+00
#> 48: -1.666683505  0.082         7--6 0.082 -1.810329e+00
#> 49: -2.662037898  0.166         7--7 0.166 -2.076102e+00
#>           log2fc pvalue LR_cell_comb p.adj            PI