exprCellCellcom

Cell-Cell communication scores based on expression only

exprCellCellcom(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  cluster_column = "cell_types",
  random_iter = 1000,
  feat_set_1,
  feat_set_2,
  log2FC_addendum = 0.1,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("feats", "cells"),
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE
)

Arguments

gobject

giotto object to use

feat_type

feature type

spat_unit

spatial unit

cluster_column

cluster column with cell type information

random_iter

number of iterations

feat_set_1

first specific feature set from feature pairs

feat_set_2

second specific feature set from feature pairs

log2FC_addendum

addendum to add when calculating log2FC

detailed

provide more detailed information (random variance and z-score)

adjust_method

which method to adjust p-values

adjust_target

adjust multiple hypotheses at the cell or feature level

set_seed

set seed for random simulations (default = TRUE)

seed_number

seed number

verbose

verbose

Value

Cell-Cell communication scores for feature pairs based on expression only

Details

Statistical framework to identify if pairs of features (such as ligand-receptor combinations) are expressed at higher levels than expected based on a reshuffled null distribution of feature expression values, without considering the spatial position of cells. More details will follow soon.

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

exprCellCellcom(g, cluster_column = "leiden_clus",
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik")
#> simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  11simulation  12simulation  13simulation  14simulation  15simulation  16simulation  17simulation  18simulation  19simulation  20simulation  21simulation  22simulation  23simulation  24simulation  25simulation  26simulation  27simulation  28simulation  29simulation  30simulation  31simulation  32simulation  33simulation  34simulation  35simulation  36simulation  37simulation  38simulation  39simulation  40simulation  41simulation  42simulation  43simulation  44simulation  45simulation  46simulation  47simulation  48simulation  49simulation  50simulation  51simulation  52simulation  53simulation  54simulation  55simulation  56simulation  57simulation  58simulation  59simulation  60simulation  61simulation  62simulation  63simulation  64simulation  65simulation  66simulation  67simulation  68simulation  69simulation  70simulation  71simulation  72simulation  73simulation  74simulation  75simulation  76simulation  77simulation  78simulation  79simulation  80simulation  81simulation  82simulation  83simulation  84simulation  85simulation  86simulation  87simulation  88simulation  89simulation  90simulation  91simulation  92simulation  93simulation  94simulation  95simulation  96simulation  97simulation  98simulation  99simulation  100simulation  101simulation  102simulation  103simulation  104simulation  105simulation  106simulation  107simulation  108simulation  109simulation  110simulation  111simulation  112simulation  113simulation  114simulation  115simulation  116simulation  117simulation  118simulation  119simulation  120simulation  121simulation  122simulation  123simulation  124simulation  125simulation  126simulation  127simulation  128simulation  129simulation  130simulation  131simulation  132simulation  133simulation  134simulation  135simulation  136simulation  137simulation  138simulation  139simulation  140simulation  141simulation  142simulation  143simulation  144simulation  145simulation  146simulation  147simulation  148simulation  149simulation  150simulation  151simulation  152simulation  153simulation  154simulation  155simulation  156simulation  157simulation  158simulation  159simulation  160simulation  161simulation  162simulation  163simulation  164simulation  165simulation  166simulation  167simulation  168simulation  169simulation  170simulation  171simulation  172simulation  173simulation  174simulation  175simulation  176simulation  177simulation  178simulation  179simulation  180simulation  181simulation  182simulation  183simulation  184simulation  185simulation  186simulation  187simulation  188simulation  189simulation  190simulation  191simulation  192simulation  193simulation  194simulation  195simulation  196simulation  197simulation  198simulation  199simulation  200simulation  201simulation  202simulation  203simulation  204simulation  205simulation  206simulation  207simulation  208simulation  209simulation  210simulation  211simulation  212simulation  213simulation  214simulation  215simulation  216simulation  217simulation  218simulation  219simulation  220simulation  221simulation  222simulation  223simulation  224simulation  225simulation  226simulation  227simulation  228simulation  229simulation  230simulation  231simulation  232simulation  233simulation  234simulation  235simulation  236simulation  237simulation  238simulation  239simulation  240simulation  241simulation  242simulation  243simulation  244simulation  245simulation  246simulation  247simulation  248simulation  249simulation  250simulation  251simulation  252simulation  253simulation  254simulation  255simulation  256simulation  257simulation  258simulation  259simulation  260simulation  261simulation  262simulation  263simulation  264simulation  265simulation  266simulation  267simulation  268simulation  269simulation  270simulation  271simulation  272simulation  273simulation  274simulation  275simulation  276simulation  277simulation  278simulation  279simulation  280simulation  281simulation  282simulation  283simulation  284simulation  285simulation  286simulation  287simulation  288simulation  289simulation  290simulation  291simulation  292simulation  293simulation  294simulation  295simulation  296simulation  297simulation  298simulation  299simulation  300simulation  301simulation  302simulation  303simulation  304simulation  305simulation  306simulation  307simulation  308simulation  309simulation  310simulation  311simulation  312simulation  313simulation  314simulation  315simulation  316simulation  317simulation  318simulation  319simulation  320simulation  321simulation  322simulation  323simulation  324simulation  325simulation  326simulation  327simulation  328simulation  329simulation  330simulation  331simulation  332simulation  333simulation  334simulation  335simulation  336simulation  337simulation  338simulation  339simulation  340simulation  341simulation  342simulation  343simulation  344simulation  345simulation  346simulation  347simulation  348simulation  349simulation  350simulation  351simulation  352simulation  353simulation  354simulation  355simulation  356simulation  357simulation  358simulation  359simulation  360simulation  361simulation  362simulation  363simulation  364simulation  365simulation  366simulation  367simulation  368simulation  369simulation  370simulation  371simulation  372simulation  373simulation  374simulation  375simulation  376simulation  377simulation  378simulation  379simulation  380simulation  381simulation  382simulation  383simulation  384simulation  385simulation  386simulation  387simulation  388simulation  389simulation  390simulation  391simulation  392simulation  393simulation  394simulation  395simulation  396simulation  397simulation  398simulation  399simulation  400simulation  401simulation  402simulation  403simulation  404simulation  405simulation  406simulation  407simulation  408simulation  409simulation  410simulation  411simulation  412simulation  413simulation  414simulation  415simulation  416simulation  417simulation  418simulation  419simulation  420simulation  421simulation  422simulation  423simulation  424simulation  425simulation  426simulation  427simulation  428simulation  429simulation  430simulation  431simulation  432simulation  433simulation  434simulation  435simulation  436simulation  437simulation  438simulation  439simulation  440simulation  441simulation  442simulation  443simulation  444simulation  445simulation  446simulation  447simulation  448simulation  449simulation  450simulation  451simulation  452simulation  453simulation  454simulation  455simulation  456simulation  457simulation  458simulation  459simulation  460simulation  461simulation  462simulation  463simulation  464simulation  465simulation  466simulation  467simulation  468simulation  469simulation  470simulation  471simulation  472simulation  473simulation  474simulation  475simulation  476simulation  477simulation  478simulation  479simulation  480simulation  481simulation  482simulation  483simulation  484simulation  485simulation  486simulation  487simulation  488simulation  489simulation  490simulation  491simulation  492simulation  493simulation  494simulation  495simulation  496simulation  497simulation  498simulation  499simulation  500simulation  501simulation  502simulation  503simulation  504simulation  505simulation  506simulation  507simulation  508simulation  509simulation  510simulation  511simulation  512simulation  513simulation  514simulation  515simulation  516simulation  517simulation  518simulation  519simulation  520simulation  521simulation  522simulation  523simulation  524simulation  525simulation  526simulation  527simulation  528simulation  529simulation  530simulation  531simulation  532simulation  533simulation  534simulation  535simulation  536simulation  537simulation  538simulation  539simulation  540simulation  541simulation  542simulation  543simulation  544simulation  545simulation  546simulation  547simulation  548simulation  549simulation  550simulation  551simulation  552simulation  553simulation  554simulation  555simulation  556simulation  557simulation  558simulation  559simulation  560simulation  561simulation  562simulation  563simulation  564simulation  565simulation  566simulation  567simulation  568simulation  569simulation  570simulation  571simulation  572simulation  573simulation  574simulation  575simulation  576simulation  577simulation  578simulation  579simulation  580simulation  581simulation  582simulation  583simulation  584simulation  585simulation  586simulation  587simulation  588simulation  589simulation  590simulation  591simulation  592simulation  593simulation  594simulation  595simulation  596simulation  597simulation  598simulation  599simulation  600simulation  601simulation  602simulation  603simulation  604simulation  605simulation  606simulation  607simulation  608simulation  609simulation  610simulation  611simulation  612simulation  613simulation  614simulation  615simulation  616simulation  617simulation  618simulation  619simulation  620simulation  621simulation  622simulation  623simulation  624simulation  625simulation  626simulation  627simulation  628simulation  629simulation  630simulation  631simulation  632simulation  633simulation  634simulation  635simulation  636simulation  637simulation  638simulation  639simulation  640simulation  641simulation  642simulation  643simulation  644simulation  645simulation  646simulation  647simulation  648simulation  649simulation  650simulation  651simulation  652simulation  653simulation  654simulation  655simulation  656simulation  657simulation  658simulation  659simulation  660simulation  661simulation  662simulation  663simulation  664simulation  665simulation  666simulation  667simulation  668simulation  669simulation  670simulation  671simulation  672simulation  673simulation  674simulation  675simulation  676simulation  677simulation  678simulation  679simulation  680simulation  681simulation  682simulation  683simulation  684simulation  685simulation  686simulation  687simulation  688simulation  689simulation  690simulation  691simulation  692simulation  693simulation  694simulation  695simulation  696simulation  697simulation  698simulation  699simulation  700simulation  701simulation  702simulation  703simulation  704simulation  705simulation  706simulation  707simulation  708simulation  709simulation  710simulation  711simulation  712simulation  713simulation  714simulation  715simulation  716simulation  717simulation  718simulation  719simulation  720simulation  721simulation  722simulation  723simulation  724simulation  725simulation  726simulation  727simulation  728simulation  729simulation  730simulation  731simulation  732simulation  733simulation  734simulation  735simulation  736simulation  737simulation  738simulation  739simulation  740simulation  741simulation  742simulation  743simulation  744simulation  745simulation  746simulation  747simulation  748simulation  749simulation  750simulation  751simulation  752simulation  753simulation  754simulation  755simulation  756simulation  757simulation  758simulation  759simulation  760simulation  761simulation  762simulation  763simulation  764simulation  765simulation  766simulation  767simulation  768simulation  769simulation  770simulation  771simulation  772simulation  773simulation  774simulation  775simulation  776simulation  777simulation  778simulation  779simulation  780simulation  781simulation  782simulation  783simulation  784simulation  785simulation  786simulation  787simulation  788simulation  789simulation  790simulation  791simulation  792simulation  793simulation  794simulation  795simulation  796simulation  797simulation  798simulation  799simulation  800simulation  801simulation  802simulation  803simulation  804simulation  805simulation  806simulation  807simulation  808simulation  809simulation  810simulation  811simulation  812simulation  813simulation  814simulation  815simulation  816simulation  817simulation  818simulation  819simulation  820simulation  821simulation  822simulation  823simulation  824simulation  825simulation  826simulation  827simulation  828simulation  829simulation  830simulation  831simulation  832simulation  833simulation  834simulation  835simulation  836simulation  837simulation  838simulation  839simulation  840simulation  841simulation  842simulation  843simulation  844simulation  845simulation  846simulation  847simulation  848simulation  849simulation  850simulation  851simulation  852simulation  853simulation  854simulation  855simulation  856simulation  857simulation  858simulation  859simulation  860simulation  861simulation  862simulation  863simulation  864simulation  865simulation  866simulation  867simulation  868simulation  869simulation  870simulation  871simulation  872simulation  873simulation  874simulation  875simulation  876simulation  877simulation  878simulation  879simulation  880simulation  881simulation  882simulation  883simulation  884simulation  885simulation  886simulation  887simulation  888simulation  889simulation  890simulation  891simulation  892simulation  893simulation  894simulation  895simulation  896simulation  897simulation  898simulation  899simulation  900simulation  901simulation  902simulation  903simulation  904simulation  905simulation  906simulation  907simulation  908simulation  909simulation  910simulation  911simulation  912simulation  913simulation  914simulation  915simulation  916simulation  917simulation  918simulation  919simulation  920simulation  921simulation  922simulation  923simulation  924simulation  925simulation  926simulation  927simulation  928simulation  929simulation  930simulation  931simulation  932simulation  933simulation  934simulation  935simulation  936simulation  937simulation  938simulation  939simulation  940simulation  941simulation  942simulation  943simulation  944simulation  945simulation  946simulation  947simulation  948simulation  949simulation  950simulation  951simulation  952simulation  953simulation  954simulation  955simulation  956simulation  957simulation  958simulation  959simulation  960simulation  961simulation  962simulation  963simulation  964simulation  965simulation  966simulation  967simulation  968simulation  969simulation  970simulation  971simulation  972simulation  973simulation  974simulation  975simulation  976simulation  977simulation  978simulation  979simulation  980simulation  981simulation  982simulation  983simulation  984simulation  985simulation  986simulation  987simulation  988simulation  989simulation  990simulation  991simulation  992simulation  993simulation  994simulation  995simulation  996simulation  997simulation  998simulation  999simulation  1000
#> Key: <LR_comb>
#>                   LR_comb lig_cell_type   lig_expr  ligand rec_cell_type
#>                    <char>        <fctr>      <num>  <char>        <fctr>
#>  1: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             5
#>  2: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             3
#>  3: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             7
#>  4: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             1
#>  5: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             6
#>  6: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             5
#>  7: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             3
#>  8: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             5
#>  9: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             2
#> 10: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             3
#> 11: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             4
#> 12: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             7
#> 13: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             1
#> 14: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             5
#> 15: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             7
#> 16: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             3
#> 17: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             6
#> 18: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             5
#> 19: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             1
#> 20: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             5
#> 21: Gm19935-9630013A20Rik             8 0.65226568 Gm19935             8
#> 22: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             3
#> 23: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             3
#> 24: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             6
#> 25: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             7
#> 26: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             2
#> 27: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             7
#> 28: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             1
#> 29: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             4
#> 30: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             2
#> 31: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             7
#> 32: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             5
#> 33: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             3
#> 34: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             6
#> 35: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             1
#> 36: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             1
#> 37: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             4
#> 38: Gm19935-9630013A20Rik             7 0.07989824 Gm19935             5
#> 39: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             6
#> 40: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             6
#> 41: Gm19935-9630013A20Rik             7 0.07989824 Gm19935             3
#> 42: Gm19935-9630013A20Rik             6 0.44158783 Gm19935             8
#> 43: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             2
#> 44: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             7
#> 45: Gm19935-9630013A20Rik             2 0.39966555 Gm19935             8
#> 46: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             2
#> 47: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             4
#> 48: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             1
#> 49: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             2
#> 50: Gm19935-9630013A20Rik             7 0.07989824 Gm19935             7
#> 51: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             6
#> 52: Gm19935-9630013A20Rik             4 0.25164714 Gm19935             4
#> 53: Gm19935-9630013A20Rik             1 0.24398768 Gm19935             4
#> 54: Gm19935-9630013A20Rik             7 0.07989824 Gm19935             1
#> 55: Gm19935-9630013A20Rik             7 0.07989824 Gm19935             6
#> 56: Gm19935-9630013A20Rik             5 0.28866359 Gm19935             8
#> 57: Gm19935-9630013A20Rik             3 0.13054788 Gm19935             2
#>      rec_expr      receptor    LR_expr lig_nr rec_nr rand_expr      av_diff
#>         <num>        <char>      <num>  <int>  <int>     <num>        <num>
#>  1: 0.4105604 9630013A20Rik 1.06282603     67     58 0.5100021  0.552823975
#>  2: 0.3944654 9630013A20Rik 1.04673103     67     96 0.5062487  0.540482355
#>  3: 0.2980556 9630013A20Rik 0.95032125     67     74 0.5170216  0.433299654
#>  4: 0.2609334 9630013A20Rik 0.91319907     67    128 0.5074076  0.405791422
#>  5: 0.2311754 9630013A20Rik 0.88344105     67      7 0.5070495  0.376391594
#>  6: 0.4105604 9630013A20Rik 0.85214818      7     58 0.5312240  0.320924156
#>  7: 0.3944654 9630013A20Rik 0.83605318      7     96 0.5274706  0.308582537
#>  8: 0.4105604 9630013A20Rik 0.81022591    151     58 0.5298379  0.280387961
#>  9: 0.1451182 9630013A20Rik 0.79738389     67    151 0.5085692  0.288814732
#> 10: 0.3944654 9630013A20Rik 0.79413090    151     96 0.5260846  0.268046342
#> 11: 0.1038294 9630013A20Rik 0.75609512     67     43 0.5058725  0.250222655
#> 12: 0.2980556 9630013A20Rik 0.73964340      7     74 0.5382436  0.201399836
#> 13: 0.2609334 9630013A20Rik 0.70252122      7    128 0.5286296  0.173891604
#> 14: 0.4105604 9630013A20Rik 0.69922395     58     58 0.5287210  0.170502986
#> 15: 0.2980556 9630013A20Rik 0.69772112    151     74 0.5368575  0.160863640
#> 16: 0.3944654 9630013A20Rik 0.68312895     58     96 0.5249676  0.158161366
#> 17: 0.2311754 9630013A20Rik 0.67276320      7      7 0.5282714  0.144491775
#> 18: 0.4105604 9630013A20Rik 0.66220749     43     58 0.5317640  0.130443499
#> 19: 0.2609334 9630013A20Rik 0.66059894    151    128 0.5272435  0.133355408
#> 20: 0.4105604 9630013A20Rik 0.65454804    128     58 0.5323289  0.122219173
#> 21: 0.0000000 9630013A20Rik 0.65226568     67     67 0.5033698  0.148895859
#> 22: 0.3944654 9630013A20Rik 0.64611249     43     96 0.5280106  0.118101879
#> 23: 0.3944654 9630013A20Rik 0.63845303    128     96 0.5285755  0.109877554
#> 24: 0.2311754 9630013A20Rik 0.63084093    151      7 0.5268854  0.103955580
#> 25: 0.2980556 9630013A20Rik 0.58671917     58     74 0.5357405  0.050978665
#> 26: 0.1451182 9630013A20Rik 0.58670604      7    151 0.5297911  0.056914914
#> 27: 0.2980556 9630013A20Rik 0.54970271     43     74 0.5387835  0.010919178
#> 28: 0.2609334 9630013A20Rik 0.54959698     58    128 0.5261266  0.023470433
#> 29: 0.1038294 9630013A20Rik 0.54541727      7     43 0.5270944  0.018322837
#> 30: 0.1451182 9630013A20Rik 0.54478377    151    151 0.5284051  0.016378718
#> 31: 0.2980556 9630013A20Rik 0.54204325    128     74 0.5393484  0.002694852
#> 32: 0.4105604 9630013A20Rik 0.54110823     96     58 0.5264708  0.014637424
#> 33: 0.3944654 9630013A20Rik 0.52501323     96     96 0.5227174  0.002295805
#> 34: 0.2311754 9630013A20Rik 0.51983897     58      7 0.5257684 -0.005929395
#> 35: 0.2609334 9630013A20Rik 0.51258052     43    128 0.5291696 -0.016589054
#> 36: 0.2609334 9630013A20Rik 0.50492107    128    128 0.5297344 -0.024813380
#> 37: 0.1038294 9630013A20Rik 0.50349499    151     43 0.5257084 -0.022213358
#> 38: 0.4105604 9630013A20Rik 0.49045859     74     58 0.5262103 -0.035751683
#> 39: 0.2311754 9630013A20Rik 0.48282251     43      7 0.5288114 -0.045988882
#> 40: 0.2311754 9630013A20Rik 0.47516306    128      7 0.5293763 -0.054213208
#> 41: 0.3944654 9630013A20Rik 0.47436359     74     96 0.5224569 -0.048093302
#> 42: 0.0000000 9630013A20Rik 0.44158783      7     67 0.5245918 -0.083003959
#> 43: 0.1451182 9630013A20Rik 0.43378181     58    151 0.5272881 -0.093506257
#> 44: 0.2980556 9630013A20Rik 0.42860345     96     74 0.5334903 -0.104886897
#> 45: 0.0000000 9630013A20Rik 0.39966555    151     67 0.5232057 -0.123540154
#> 46: 0.1451182 9630013A20Rik 0.39676535     43    151 0.5303311 -0.133565744
#> 47: 0.1038294 9630013A20Rik 0.39249303     58     43 0.5245914 -0.132098334
#> 48: 0.2609334 9630013A20Rik 0.39148127     96    128 0.5238764 -0.132395129
#> 49: 0.1451182 9630013A20Rik 0.38910590    128    151 0.5308960 -0.141790070
#> 50: 0.2980556 9630013A20Rik 0.37795381     74     74 0.5332298 -0.155276004
#> 51: 0.2311754 9630013A20Rik 0.36172325     96      7 0.5235182 -0.161794957
#> 52: 0.1038294 9630013A20Rik 0.35547658     43     43 0.5276344 -0.172157821
#> 53: 0.1038294 9630013A20Rik 0.34781712    128     43 0.5281993 -0.180382146
#> 54: 0.2609334 9630013A20Rik 0.34083163     74    128 0.5236159 -0.182784236
#> 55: 0.2311754 9630013A20Rik 0.31107362     74      7 0.5232577 -0.212184064
#> 56: 0.0000000 9630013A20Rik 0.28866359     58     67 0.5220887 -0.233425130
#> 57: 0.1451182 9630013A20Rik 0.27566609     96    151 0.5250379 -0.249371819
#>           log2fc pvalue LR_cell_comb p.adj            PI
#>            <num>  <num>       <char> <num>         <num>
#>  1:  0.930749258  0.156         8--5 0.156  0.7509986818
#>  2:  0.919545451  0.150         8--3 0.150  0.7576215345
#>  3:  0.767437767  0.246         8--7 0.246  0.4674194015
#>  4:  0.738180683  0.250         8--1 0.250  0.4444290553
#>  5:  0.696024514  0.288         8--6 0.288  0.3762760812
#>  6:  0.593033997  0.044         6--5 0.044  0.8044786813
#>  7:  0.577042536  0.040         6--3 0.040  0.8066708476
#>  8:  0.531243968  0.036         2--5 0.036  0.7669555887
#>  9:  0.560304062  0.412         8--2 0.412  0.2157746542
#> 10:  0.514128534  0.020         2--3 0.020  0.8734889584
#> 11:  0.498756964  0.448         8--4 0.448  0.1739275190
#> 12:  0.395669655  0.258         6--7 0.258  0.2328042279
#> 13:  0.352329299  0.210         6--1 0.210  0.2388020010
#> 14:  0.346179946  0.250         5--5 0.250  0.2084210952
#> 15:  0.324913920  0.306         2--7 0.306  0.1670962672
#> 16:  0.325468523  0.234         5--3 0.234  0.2053003830
#> 17:  0.298638427  0.386         6--6 0.386  0.1234609170
#> 18:  0.270798074  0.364         4--5 0.364  0.1188528999
#> 19:  0.278110239  0.190         2--1 0.190  0.2005860086
#> 20:  0.254937674  0.318         1--5 0.318  0.1268500727
#> 21:  0.318199737  0.650         8--8 0.650  0.0595309207
#> 22:  0.248604226  0.372         4--3 0.372  0.1067648400
#> 23:  0.232420173  0.304         1--3 0.304  0.1201906109
#> 24:  0.221355818  0.400         2--6 0.400  0.0880863361
#> 25:  0.111282228  0.718         5--7 0.718  0.0160107924
#> 26:  0.124819222  0.636         6--2 0.636  0.0245323300
#> 27:  0.024452603  0.908         4--7 0.908  0.0010249101
#> 28:  0.053090654  0.832         5--1 0.832  0.0042407048
#> 29:  0.041549470  0.832         6--4 0.832  0.0033188335
#> 30:  0.037120652  0.876         2--2 0.876  0.0021342851
#> 31:  0.006068177  0.932         1--7 0.932  0.0001855897
#> 32:  0.033320648  0.846         3--5 0.846  0.0024200665
#> 33:  0.005309078  0.948         3--3 0.948  0.0001231263
#> 34: -0.013735265  0.968         5--6 0.968 -0.0001940057
#> 35: -0.038549415  0.918         4--1 0.918 -0.0014323929
#> 36: -0.057996680  0.824         1--1 0.824 -0.0048759426
#> 37: -0.052148558  0.866         2--4 0.866 -0.0032583518
#> 38: -0.084811291  0.862         7--5 0.862 -0.0054697120
#> 39: -0.109570756  0.796         4--6 0.796 -0.0108570301
#> 40: -0.129951758  0.714         1--6 0.714 -0.0190121746
#> 41: -0.116009631  0.772         7--3 0.772 -0.0130374755
#> 42: -0.205718284  0.948         6--8 0.948 -0.0047709490
#> 43: -0.232877976  0.502         5--2 0.502 -0.0696995125
#> 44: -0.261136796  0.548         3--7 0.548 -0.0682140080
#> 45: -0.318745683  0.928         2--8 0.928 -0.0103439425
#> 46: -0.343545284  0.250         4--2 0.250 -0.2068348707
#> 47: -0.342809311  0.380         5--4 0.380 -0.1440540957
#> 48: -0.344123799  0.202         3--1 0.202 -0.2390451258
#> 49: -0.367255269  0.132         1--2 0.132 -0.3229738575
#> 50: -0.405857981  0.434         7--7 0.434 -0.1471276864
#> 51: -0.433403304  0.238         3--6 0.238 -0.2701936066
#> 52: -0.462547557  0.238         4--4 0.238 -0.2883628056
#> 53: -0.488312573  0.110         1--4 0.110 -0.4681000047
#> 54: -0.500429894  0.224         7--1 0.224 -0.3251553151
#> 55: -0.600431977  0.240         7--6 0.240 -0.3721409895
#> 56: -0.678598377  0.622         5--8 0.622 -0.1399335102
#> 57: -0.734492795  0.018         3--2 0.018 -1.2814897737
#>  [ reached getOption("max.print") -- omitted 8 rows ]