findGiniMarkers

Identify marker feats for selected clusters based on gini detection and expression scores.

findGiniMarkers(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  group_1 = NULL,
  group_1_name = NULL,
  group_2 = NULL,
  group_2_name = NULL,
  min_expr_gini_score = 0.2,
  min_det_gini_score = 0.2,
  detection_threshold = 0,
  rank_score = 1,
  min_feats = 5,
  min_genes = NULL
)

Arguments

gobject

giotto object

feat_type

feature type

spat_unit

spatial unit

expression_values

feat expression values to use

cluster_column

clusters to use

subset_clusters

selection of clusters to compare

group_1

group 1 cluster IDs from cluster_column for pairwise comparison

group_1_name

custom name for group_1 clusters

group_2

group 2 cluster IDs from cluster_column for pairwise comparison

group_2_name

custom name for group_2 clusters

min_expr_gini_score

filter on minimum gini coefficient for expression

min_det_gini_score

filter on minimum gini coefficient for detection

detection_threshold

detection threshold for feat expression

rank_score

rank scores for both detection and expression to include

min_feats

minimum number of top feats to return

min_genes

deprecated, use min_feats

Value

data.table with marker feats

Details

Detection of marker feats using the gini coefficient is based on the following steps/principles per feat:

calculate average expression per cluster

calculate detection fraction per cluster

calculate gini-coefficient for av. expression values over all clusters

calculate gini-coefficient for detection fractions over all clusters

convert gini-scores to rank scores

for each feat create combined score = detection rank x expression rank x expr gini-coefficient x detection gini-coefficient

for each feat sort on expression and detection rank and combined score

As a results "top gini" feats are feats that are very selectivily expressed in a specific cluster, however not always expressed in all cells of that cluster. In other words highly specific, but not necessarily sensitive at the single-cell level.

To perform differential expression between custom selected groups of cells you need to specify the cell_ID column to parameter cluster_column and provide the individual cell IDs to the parameters group_1 and group_2

By default group names will be created by pasting the different id names within each selected group. When you have many different ids in a single group it is recommend to provide names for both groups to group_1_name and group_2_name

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

findGiniMarkers(g, cluster_column = "leiden_clus")
#>               feats cluster expression expression_gini detection detection_gini
#>              <char>  <char>      <num>           <num>     <num>          <num>
#>    1: 1700001C02Rik       2  0.1803989      0.44884761 0.0546875    0.380216472
#>    2:         Crygn       2  0.1604868      0.47269447 0.0546875    0.429912372
#>    3:        Mustn1       2  0.2550445      0.38682356 0.0937500    0.389320687
#>    4:        Ccdc33       2  0.1491296      0.54325707 0.0468750    0.483041252
#>    5:         Odf3b       2  0.1855937      0.34524480 0.0703125    0.366683556
#>   ---                                                                          
#> 3982:          Ngef       5  2.5421599      0.04052548 0.6417910    0.030121744
#> 3983:          Chgb       5  4.6218517      0.02970958 0.8656716    0.029415603
#> 3984:          Apod       5  4.6848957      0.02809576 0.9253731    0.012387359
#> 3985:          Mobp       5  6.4262032      0.01278243 0.9850746    0.003813677
#> 3986:           Ddn       5  4.7020073      0.01347302 0.8507463    0.017656944
#>       expression_rank detection_rank   comb_score comb_rank
#>                 <num>          <num>        <num>     <int>
#>    1:       1.0000000      1.0000000 1.706593e-01         1
#>    2:       0.8714286      0.8714286 1.543207e-01         2
#>    3:       1.0000000      1.0000000 1.505984e-01         3
#>    4:       0.7428571      0.7428571 1.448106e-01         4
#>    5:       1.0000000      1.0000000 1.265956e-01         5
#>   ---                                                      
#> 3982:       0.1000000      0.1000000 1.220698e-05       630
#> 3983:       0.1000000      0.1000000 8.739253e-06       631
#> 3984:       0.2285714      0.1000000 7.955022e-06       632
#> 3985:       0.4857143      0.2285714 5.412028e-06       633
#> 3986:       0.1000000      0.1000000 2.378925e-06       634