findMarkers

Identify marker feats for selected clusters.

findMarkers(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column = NULL,
  method = c("scran", "gini", "mast"),
  subset_clusters = NULL,
  group_1 = NULL,
  group_2 = NULL,
  min_expr_gini_score = 0.5,
  min_det_gini_score = 0.5,
  detection_threshold = 0,
  rank_score = 1,
  min_feats = 4,
  min_genes = NULL,
  group_1_name = NULL,
  group_2_name = NULL,
  adjust_columns = NULL,
  ...
)

Arguments

gobject

giotto object

spat_unit

spatial unit

feat_type

feature type

expression_values

feat expression values to use

cluster_column

clusters to use

method

method to use to detect differentially expressed feats

subset_clusters

selection of clusters to compare

group_1

group 1 cluster IDs from cluster_column for pairwise comparison

group_2

group 2 cluster IDs from cluster_column for pairwise comparison

min_expr_gini_score

gini: filter on minimum gini coefficient for expression

min_det_gini_score

gini: filter minimum gini coefficient for detection

detection_threshold

gini: detection threshold for feat expression

rank_score

gini: rank scores to include

min_feats

minimum number of top feats to return (for gini)

min_genes

deprecated, use min_feats

group_1_name

mast: custom name for group_1 clusters

group_2_name

mast: custom name for group_2 clusters

adjust_columns

mast: column in pDataDT to adjust for (e.g. detection rate)

...

additional parameters for the findMarkers function in scran or zlm function in MAST

Value

data.table with marker feats

Details

Wrapper for all individual functions to detect marker feats for clusters.

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

findMarkers(g, cluster_column = "leiden_clus")
#> Using 'Scran' to detect marker genes. If used in published 
#>         research, please cite:
#>         Lun ATL, McCarthy DJ, Marioni JC (2016).
#>         'A step-by-step workflow for low-level analysis of single-cell RNA-seq 
#>         data with Bioconductor.' 
#>         F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2.
#> [[1]]
#>        Top      p.value          FDR summary.logFC      logFC.2      logFC.3
#>      <int>        <num>        <num>         <num>        <num>        <num>
#>   1:     1 1.035419e-02 1.025712e-01    -0.7424665 -0.345955606 -0.742466517
#>   2:     1 2.818778e-03 3.504128e-02     0.2214652  0.045814646  0.062023374
#>   3:     1 1.220969e-02 1.190914e-01     0.5397091  0.014622023  0.070384982
#>   4:     1 2.381963e-03 3.213117e-02     0.9058808  0.531787411  0.029978539
#>   5:     1 3.318565e-21 2.103970e-18     1.6846287  0.099009163  0.069509380
#>  ---                                                                        
#> 630:   316 9.852089e-01 9.935011e-01     0.1663032  0.023042514 -0.026761247
#> 631:   317 8.938272e-01 9.915093e-01    -0.5447480 -0.101469788  0.036607054
#> 632:   347 8.844787e-01 9.915093e-01    -0.1446356  0.065406419 -0.144635624
#> 633:   347 9.457384e-01 9.915093e-01    -0.1842519  0.049415938  0.067821739
#> 634:   351 9.676506e-01 9.915093e-01     0.1977989 -0.005655565  0.006412597
#>           logFC.4     logFC.5      logFC.6      logFC.7    logFC.8   feats
#>             <num>       <num>        <num>        <num>      <num>  <char>
#>   1: -0.097231029 -0.80525591 -0.231563731 -0.496597553 -0.4318011    Cygb
#>   2:  0.221465223  0.01404530 -0.007627079 -0.134873150  0.2214652    Cd74
#>   3:  0.128413596  0.53970915 -0.370912513  0.239238103  0.6473761   Pde8a
#>   4:  0.526967687  0.03199585  0.905880797 -0.138322635  1.2444069   Cnih3
#>   5:  0.184042749 -0.12744366  0.127781888 -0.105845753  1.6846287 Tmem63a
#>  ---                                                                      
#> 630:  0.004134948  0.09454750  0.015091475  0.166303237  0.2485584    Grm4
#> 631:  0.122741310  0.10876869  0.034423186  0.053039480 -0.5447480     C4b
#> 632: -0.084747461 -0.04184853  0.155082104  0.190451013  0.2284142   Cers2
#> 633: -0.148703774 -0.01850428 -0.066323851 -0.184251893  0.1198168    Klk8
#> 634: -0.059335133 -0.08914093 -0.062026513 -0.009840105  0.1977989   Myo1d
#>      cluster
#>       <char>
#>   1:       1
#>   2:       1
#>   3:       1
#>   4:       1
#>   5:       1
#>  ---        
#> 630:       1
#> 631:       1
#> 632:       1
#> 633:       1
#> 634:       1
#> 
#> [[2]]
#>        Top      p.value          FDR summary.logFC      logFC.1     logFC.3
#>      <int>        <num>        <num>         <num>        <num>       <num>
#>   1:     1 1.322858e-16 5.402207e-14     1.5856196 -0.099009163 -0.02949978
#>   2:     1 8.773095e-03 9.118265e-02     0.4415731  0.096853187  0.44157306
#>   3:     1 7.039234e-10 2.789296e-08     0.9821925  0.268279231  0.21709059
#>   4:     1 1.988321e-09 7.415269e-08     0.8085488  0.234184924  0.01202878
#>   5:     1 1.701503e-10 8.298100e-09     0.9651944  0.251113227  0.21319091
#>  ---                                                                       
#> 630:   317 8.808135e-01 9.988226e-01    -0.1082076 -0.108207647 -0.10067873
#> 631:   320 9.925209e-01 9.988226e-01     0.1454975  0.145497550 -0.05277827
#> 632:   349 9.540091e-01 9.988226e-01    -0.1002898 -0.100289836  0.07368298
#> 633:   377 9.850613e-01 9.988226e-01     0.2034544  0.005655565  0.01206816
#> 634:   381 9.741021e-01 9.988226e-01     0.1432607 -0.023042514 -0.04980376
#>           logFC.4     logFC.5      logFC.6     logFC.7     logFC.8   feats
#>             <num>       <num>        <num>       <num>       <num>  <char>
#>   1:  0.085033587 -0.22645282  0.028772725 -0.20485492  1.58561956 Tmem63a
#>   2:  0.237632979  0.40648236  0.039075478  0.40225868 -0.04015420  Col1a2
#>   3:  0.674487431  0.30462496  0.243362188  0.36653298  0.98219250    Vav3
#>   4:  0.187501591  0.40396324  0.097382540  0.74845671  0.80854883   Cd24a
#>   5:  0.211472145  0.28299581  0.663144165 -0.15515295  0.96519441   Cpne8
#>  ---                                                                      
#> 630:  0.141969900 -0.06287110 -0.130486623  0.09215767 -0.06601173 Plekhb1
#> 631:  0.002663099 -0.07979464  0.046549063  0.01625747 -0.07372290   Olfm2
#> 632:  0.012368363  0.05912384  0.031047823  0.02075793  0.25422872    Reln
#> 633: -0.053679568 -0.08348537 -0.056370948 -0.00418454  0.20345445   Myo1d
#> 634: -0.018907566  0.07150499 -0.007951039  0.14326072  0.22551586    Grm4
#>      cluster
#>       <char>
#>   1:       2
#>   2:       2
#>   3:       2
#>   4:       2
#>   5:       2
#>  ---        
#> 630:       2
#> 631:       2
#> 632:       2
#> 633:       2
#> 634:       2
#> 
#> [[3]]
#>        Top      p.value          FDR summary.logFC      logFC.1     logFC.2
#>      <int>        <num>        <num>         <num>        <num>       <num>
#>   1:     1 1.035419e-02 1.367616e-01    0.74246652  0.742466517  0.39651091
#>   2:     1 6.234905e-03 8.983931e-02    0.89463370  0.491402552  0.36062313
#>   3:     1 2.734616e-14 1.733746e-11    1.61511934 -0.069509380  0.02949978
#>   4:     1 2.198645e-03 3.574208e-02    1.14026764  0.499405006  0.61920164
#>   5:     1 8.773095e-03 1.183434e-01   -0.44157306 -0.344719873 -0.44157306
#>  ---                                                                       
#> 630:   321 8.725045e-01 9.944099e-01    0.22332696  0.138780968 -0.08311886
#> 631:   323 8.082689e-01 9.944099e-01   -0.24639347 -0.078965586  0.06476361
#> 632:   337 8.996667e-01 9.944099e-01   -0.18810973  0.033535710  0.03669351
#> 633:   373 9.645596e-01 9.944099e-01   -0.15497403  0.040762316 -0.07992615
#> 634:   376 9.655713e-01 9.944099e-01   -0.09555353 -0.006412597 -0.01206816
#>          logFC.4     logFC.5     logFC.6     logFC.7     logFC.8   feats
#>            <num>       <num>       <num>       <num>       <num>  <char>
#>   1:  0.64523549 -0.06278939  0.51090279  0.24586896  0.31066542    Cygb
#>   2:  0.89463370 -0.05318974  0.04480022  0.66141269  0.39497336 Serinc5
#>   3:  0.11453337 -0.19695304  0.05827251 -0.17535513  1.61511934 Tmem63a
#>   4:  0.37116792  0.76857021  1.14026764  0.54128224  1.11382037   Lamp5
#>   5: -0.20394008 -0.03509070 -0.40249758 -0.03931438 -0.48172726  Col1a2
#>  ---                                                                    
#> 630:  0.14127325  0.08979251  0.22332696  0.03987848  0.15428540  Cldn11
#> 631:  0.15398391 -0.07190171 -0.24639347  0.12646963 -0.69153821   Cpne4
#> 632:  0.05227291 -0.18810973  0.06549068 -0.19324861 -0.52685495  Necab2
#> 633: -0.15497403  0.11304734 -0.11120988 -0.03660581 -0.03241915  Cx3cr1
#> 634: -0.06574773 -0.09555353 -0.06843911 -0.01625270  0.19138629   Myo1d
#>      cluster
#>       <char>
#>   1:       3
#>   2:       3
#>   3:       3
#>   4:       3
#>   5:       3
#>  ---        
#> 630:       3
#> 631:       3
#> 632:       3
#> 633:       3
#> 634:       3
#> 
#> [[4]]
#>        Top      p.value          FDR summary.logFC     logFC.1      logFC.2
#>      <int>        <num>        <num>         <num>       <num>        <num>
#>   1:     1 4.078338e-02 3.605956e-01    -0.8194164 -0.37140326 -0.296402751
#>   2:     1 6.234905e-03 1.166474e-01    -0.8946337 -0.40323115 -0.534010573
#>   3:     1 5.637547e-03 1.116939e-01    -0.2214652 -0.22146522 -0.175650578
#>   4:     1 9.765949e-04 2.579838e-02    -0.6744874 -0.40620820 -0.674487431
#>   5:     1 3.355844e-10 2.127605e-07     1.6132097 -0.03224207  0.240097326
#>  ---                                                                       
#> 630:   368 8.208759e-01 9.983622e-01    -0.1803725 -0.08898414  0.061729404
#> 631:   373 9.550395e-01 9.983622e-01     0.1420756  0.11308022  0.142075559
#> 632:   393 9.141683e-01 9.983622e-01    -0.1886204 -0.09799835  0.140919639
#> 633:   394 9.768809e-01 9.983622e-01    -0.1126582 -0.11265820 -0.012368363
#> 634:   398 9.925209e-01 9.983622e-01     0.1428345  0.14283445 -0.002663099
#>          logFC.3     logFC.5     logFC.6      logFC.7    logFC.8   feats
#>            <num>       <num>       <num>        <num>      <num>  <char>
#>   1: -0.37471700 -0.07852149 -0.81941640  0.059540141 -0.2958488   Itih3
#>   2: -0.89463370 -0.94782344 -0.84983348 -0.233221005 -0.4996603 Serinc5
#>   3: -0.15944185 -0.20741992 -0.22909230 -0.356338373  0.0000000    Cd74
#>   4: -0.45739684 -0.36986247 -0.43112524 -0.307954446  0.3077051    Vav3
#>   5:  0.42836885  0.24410181  0.29154878  0.115797659  1.6132097  Igfbp2
#>  ---                                                                    
#> 630: -0.14261005 -0.18037248 -0.06964954  0.144637389  0.3655289   Zfhx3
#> 631: -0.11357434 -0.05007906  0.01497664 -0.078239584 -0.3651589   Htr2c
#> 632: -0.02950450  0.13664540 -0.18862038  0.099916492  0.1101629  Lgals1
#> 633:  0.06131462  0.04675548  0.01867946  0.008389562  0.2418604    Reln
#> 634: -0.05544137 -0.08245774  0.04388596  0.013594373 -0.0763860   Olfm2
#>      cluster
#>       <char>
#>   1:       4
#>   2:       4
#>   3:       4
#>   4:       4
#>   5:       4
#>  ---        
#> 630:       4
#> 631:       4
#> 632:       4
#> 633:       4
#> 634:       4
#> 
#> [[5]]
#>        Top      p.value          FDR summary.logFC     logFC.1     logFC.2
#>      <int>        <num>        <num>         <num>       <num>       <num>
#>   1:     1 3.148735e-03 7.393697e-02    -0.9106217 -0.53970915 -0.52508713
#>   2:     1 6.008411e-11 3.809332e-08     1.8120724  0.12744366  0.22645282
#>   3:     1 2.530746e-08 8.022465e-06     1.3976531  0.17892399  0.28000140
#>   4:     1 5.272255e-03 1.078261e-01    -0.6504033 -0.24658500 -0.45075136
#>   5:     1 9.508972e-03 1.674636e-01    -0.8753412 -0.48570139 -0.55316745
#>  ---                                                                      
#> 630:   328 8.582272e-01 9.998669e-01    -0.5228928 -0.08228802 -0.16108839
#> 631:   343 9.834866e-01 9.998669e-01     0.1298929 -0.04009977 -0.01678335
#> 632:   345 9.236707e-01 9.998669e-01     0.2308439  0.07024869 -0.17181603
#> 633:   367 8.069359e-01 9.998669e-01    -0.1213087 -0.09454750 -0.07150499
#> 634:   401 9.457384e-01 9.998669e-01    -0.1657476  0.01850428  0.06792022
#>          logFC.3     logFC.4      logFC.6     logFC.7    logFC.8   feats
#>            <num>       <num>        <num>       <num>      <num>  <char>
#>   1: -0.46932417 -0.41129555 -0.910621663 -0.30047105  0.1076670   Pde8a
#>   2:  0.19695304  0.31148641  0.255225548  0.02159791  1.8120724 Tmem63a
#>   3:  0.24573733 -0.01627454  0.504509233  0.91997659  1.3976531   Cd59a
#>   4: -0.65040327 -0.30694017 -0.010692430 -0.14241315  0.3563084  Galnt6
#>   5: -0.55702019 -0.87534123 -0.437304722 -0.48964771 -0.9755172  Fcer1g
#>  ---                                                                    
#> 630:  0.14550836  0.18006844  0.057070404  0.16815410 -0.5228928 Plekhh1
#> 631:  0.12989289 -0.12360199  0.005661548  0.09192136 -0.1353584   Phkg1
#> 632:  0.03875464  0.23084385  0.169667789 -0.03930303  0.1613535     Cd9
#> 633: -0.12130875 -0.09041255 -0.079456025  0.07175574  0.1540109    Grm4
#> 634:  0.08632602 -0.13019949 -0.047819566 -0.16574761  0.1383210    Klk8
#>      cluster
#>       <char>
#>   1:       5
#>   2:       5
#>   3:       5
#>   4:       5
#>   5:       5
#>  ---        
#> 630:       5
#> 631:       5
#> 632:       5
#> 633:       5
#> 634:       5
#> 
#> [[6]]
#>        Top      p.value          FDR summary.logFC     logFC.1      logFC.2
#>      <int>        <num>        <num>         <num>       <num>        <num>
#>   1:     1 3.423135e-02 0.3646465195     0.8194164  0.44801314  0.523013652
#>   2:     1 3.148735e-03 0.0767806987     0.9106217  0.37091251  0.385534536
#>   3:     1 2.381963e-03 0.0629235329    -0.9058808 -0.90588080 -0.374093386
#>   4:     1 2.052960e-07 0.0001301577     1.5568468 -0.12778189 -0.028772725
#>   5:     1 2.198645e-03 0.0629235329    -1.1402676 -0.64086263 -0.521065999
#>  ---                                                                       
#> 630:   413 9.053068e-01 0.9988914305    -0.1551640 -0.15516402  0.090779626
#> 631:   417 9.741021e-01 0.9988914305     0.1512118 -0.01509148  0.007951039
#> 632:   419 9.834866e-01 0.9988914305     0.1242313 -0.04576132 -0.022444901
#> 633:   440 8.849151e-01 0.9988914305     0.1341456  0.06632385  0.115739790
#> 634:   445 9.420166e-01 0.9988914305    -0.1263437  0.09894849 -0.046549063
#>          logFC.3     logFC.4      logFC.5     logFC.7     logFC.8   feats
#>            <num>       <num>        <num>       <num>       <num>  <char>
#>   1:  0.44469940  0.81941640  0.740894917  0.87895654  0.52356761   Itih3
#>   2:  0.44129749  0.49932611  0.910621663  0.61015062  1.01828865   Pde8a
#>   3: -0.87590226 -0.37891311 -0.873884950 -1.04420343  0.33852614   Cnih3
#>   4: -0.05827251  0.05626086 -0.255225548 -0.23362764  1.55684683 Tmem63a
#>   5: -1.14026764 -0.76909972 -0.371697425 -0.59898540 -0.02644727   Lamp5
#>  ---                                                                     
#> 630:  0.13999854  0.04324500 -0.059971263 -0.07881012 -0.27811387   Cpne7
#> 631: -0.04185272 -0.01095653  0.079456025  0.15121176  0.23346690    Grm4
#> 632:  0.12423134 -0.12926354 -0.005661548  0.08625982 -0.14101994   Phkg1
#> 633:  0.13414559 -0.08237992  0.047819566 -0.11792804  0.18614061    Klk8
#> 634: -0.09932733 -0.04388596 -0.126343708 -0.03029159 -0.12027197   Olfm2
#>      cluster
#>       <char>
#>   1:       6
#>   2:       6
#>   3:       6
#>   4:       6
#>   5:       6
#>  ---        
#> 630:       6
#> 631:       6
#> 632:       6
#> 633:       6
#> 634:       6
#> 
#> [[7]]
#>        Top      p.value          FDR summary.logFC      logFC.1     logFC.2
#>      <int>        <num>        <num>         <num>        <num>       <num>
#>   1:     1 4.136833e-08 2.622752e-05     1.7904745  0.105845753  0.20485492
#>   2:     1 2.685362e-03 1.309630e-01    -0.9362511 -0.741052608 -0.63997519
#>   3:     1 8.645155e-03 2.491377e-01     1.1327020  0.327283330  0.38512902
#>   4:     1 4.713216e-07 1.494089e-04    -0.7484567 -0.514271784 -0.74845671
#>   5:     2 1.589473e-02 3.599021e-01    -1.0071882 -0.459091477 -0.88392151
#>  ---                                                                       
#> 630:   385 9.639291e-01 9.998669e-01     0.1985063 -0.184935303  0.09398766
#> 631:   393 9.850613e-01 9.998669e-01     0.2076390  0.009840105  0.00418454
#> 632:   393 9.768809e-01 9.998669e-01    -0.1210478 -0.121047761 -0.02075793
#> 633:   401 8.948880e-01 9.998669e-01     0.1834485  0.098902487 -0.12299734
#> 634:   453 9.723458e-01 9.998669e-01     0.1292401  0.129240077 -0.01625747
#>          logFC.3      logFC.4     logFC.5     logFC.6     logFC.8   feats
#>            <num>        <num>       <num>       <num>       <num>  <char>
#>   1:  0.17535513  0.289888503 -0.02159791  0.23362764  1.79047447 Tmem63a
#>   2: -0.67423927 -0.936251137 -0.91997659 -0.41546736  0.47767655   Cd59a
#>   3:  0.27993780  0.632074178  0.43381269  1.13270198  0.11820174   Zmat4
#>   4: -0.73642793 -0.560955117 -0.34449347 -0.65107417  0.06009212   Cd24a
#>   5: -1.00718816 -0.254760122 -0.74232989 -0.85289300 -0.61100552    Gltp
#>  ---                                                                     
#> 630:  0.01883718  0.198506265  0.10980041 -0.09329166  0.20778368    Rora
#> 631:  0.01625270 -0.049495027 -0.07930083 -0.05218641  0.20763899   Myo1d
#> 632:  0.05292505 -0.008389562  0.03836592  0.01028990  0.23347080    Reln
#> 633: -0.03987848  0.101394765  0.04991403  0.18344848  0.11440692  Cldn11
#> 634: -0.06903574 -0.013594373 -0.09605212  0.03029159 -0.08998038   Olfm2
#>      cluster
#>       <char>
#>   1:       7
#>   2:       7
#>   3:       7
#>   4:       7
#>   5:       7
#>  ---        
#> 630:       7
#> 631:       7
#> 632:       7
#> 633:       7
#> 634:       7
#> 
#> [[8]]
#>        Top      p.value          FDR summary.logFC     logFC.1     logFC.2
#>      <int>        <num>        <num>         <num>       <num>       <num>
#>   1:     1 3.318565e-21 2.103970e-18    -1.6846287 -1.68462872 -1.58561956
#>   2:     1 1.345755e-20 4.266044e-18    -1.6454517 -1.64545172 -1.37311233
#>   3:     2 1.199697e-15 1.521216e-13    -1.2187292 -1.21872916 -1.11765174
#>   4:     2 1.704166e-16 3.136604e-14    -1.5470753 -1.36929216 -1.54707525
#>   5:     3 1.254111e-14 9.938833e-13    -1.1956934 -1.19569345 -1.24853119
#>  ---                                                                      
#> 630:   494 9.731903e-01 9.993036e-01    -0.2052295 -0.20522954  0.04471709
#> 631:   496 9.800557e-01 9.993036e-01    -0.2907379 -0.06299768 -0.07568060
#> 632:   511 9.917588e-01 9.993036e-01     0.2369060  0.08856425  0.13921174
#> 633:   523 9.955443e-01 9.993036e-01     0.1454665  0.07318147 -0.04750700
#> 634:   529 9.957731e-01 9.993036e-01     0.2192205  0.21922045  0.07372290
#>          logFC.3     logFC.4      logFC.5     logFC.6      logFC.7   feats
#>            <num>       <num>        <num>       <num>        <num>  <char>
#>   1: -1.61511934 -1.50058597 -1.812072380 -1.55684683 -1.790474473 Tmem63a
#>   2: -1.18484081 -1.61320965 -1.369107844 -1.32166088 -1.497411995  Igfbp2
#>   3: -1.15191582 -1.41392769 -1.397653146 -0.89314391 -0.477676551   Cd59a
#>   4: -1.39773478 -1.27405492 -1.463682872 -1.24416315 -1.294119384    Gjc2
#>   5: -1.12549696 -1.23326513 -1.381258941 -1.11735152 -1.152305989   Rgs12
#>  ---                                                                      
#> 630: -0.11399648  0.04737156 -0.042730166  0.05911527  0.026507775    Gab1
#> 631:  0.07333638  0.27577136 -0.204863136  0.02573513 -0.290737895   Ptpn3
#> 632:  0.01044922 -0.17600708  0.236905967  0.07662577  0.181992276     Ddn
#> 633:  0.03241915 -0.12255488  0.145466486 -0.07879073 -0.004186664  Cx3cr1
#> 634:  0.02094464  0.07638600 -0.006071742  0.12027197  0.089980376   Olfm2
#>      cluster
#>       <char>
#>   1:       8
#>   2:       8
#>   3:       8
#>   4:       8
#>   5:       8
#>  ---        
#> 630:       8
#> 631:       8
#> 632:       8
#> 633:       8
#> 634:       8
#>