findMarkers_one_vs_all

Identify marker feats for all clusters in a one vs all manner.

findMarkers_one_vs_all(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  method = c("scran", "gini", "mast"),
  pval = 0.01,
  logFC = 0.5,
  min_feats = 10,
  min_genes = NULL,
  min_expr_gini_score = 0.5,
  min_det_gini_score = 0.5,
  detection_threshold = 0,
  rank_score = 1,
  adjust_columns = NULL,
  verbose = TRUE,
  ...
)

Arguments

gobject

giotto object

feat_type

feature type

spat_unit

spatial unit

expression_values

feat expression values to use

cluster_column

clusters to use

subset_clusters

selection of clusters to compare

method

method to use to detect differentially expressed feats

pval

scran & mast: filter on minimal p-value

logFC

scan & mast: filter on logFC

min_feats

minimum feats to keep per cluster, overrides pval and logFC

min_genes

deprecated, use min_feats

min_expr_gini_score

gini: filter on minimum gini coefficient for expression

min_det_gini_score

gini: filter minimum gini coefficient for detection

detection_threshold

gini: detection threshold for feat expression

rank_score

gini: rank scores to include

adjust_columns

mast: column in pDataDT to adjust for (e.g. detection rate)

verbose

be verbose

...

additional parameters for the findMarkers function in scran or zlm function in MAST

Value

data.table with marker feats

Details

Wrapper for all one vs all functions to detect marker feats for clusters.

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> 3.2 read Giotto spatial centroid information
#> 3.3 read Giotto spatial overlap information
#> 4. read Giotto image information
#> 
#> checking default envname 'giotto_env'
#> a system default python environment was found
#> Using python path:
#>  "/usr/bin/python3"

findMarkers_one_vs_all(g, cluster_column = "leiden_clus")
#> using 'Scran' to detect marker feats. If used in published
#>       research, please cite: Lun ATL, McCarthy DJ, Marioni JC (2016).
#>       'A step-by-step workflow for low-level analysis of single-cell RNA-seq
#>       data with Bioconductor.'
#>       F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2. 
#> start with cluster  1start with cluster  2start with cluster  3start with cluster  4start with cluster  5start with cluster  6start with cluster  7
#>       Top     p.value        FDR     logFC    feats cluster ranking
#>     <int>       <num>      <num>     <num>   <char>  <char>   <num>
#>  1:     1 0.000354327 0.19546782 0.4994571   Arpp19       1       3
#>  2:     4 0.002932856 0.44588387 0.5955190    Pde1a       1       1
#>  3:     5 0.003516434 0.44588387 0.5293216   Ipcef1       1       2
#>  4:     6 0.008882113 0.71268782 0.4552590   Camk2d       1       6
#>  5:     7 0.013139650 0.71268782 0.4803266     Ngef       1       4
#>  6:    10 0.016171727 0.71268782 0.4738020     Tbr1       1       5
#>  7:    15 0.018401100 0.71268782 0.4359583     Myl4       1       8
#>  8:    16 0.020925524 0.71268782 0.4179815    Nptxr       1      10
#>  9:    17 0.021056054 0.71268782 0.4410319    Lamp5       1       7
#> 10:    33 0.041136065 0.76110482 0.4264191    Stx1a       1       9
#> 11:     1 0.002208825 0.53711117 0.5832163    Prr18       2       1
#> 12:     2 0.003332138 0.53711117 0.5181029  Carhsp1       2       2
#> 13:     3 0.004700090 0.53711117 0.4810502     Tprn       2       7
#> 14:     4 0.004921900 0.53711117 0.4670464    Padi2       2       9
#> 15:     5 0.005490721 0.53711117 0.4393166    S1pr5       2      11
#> 16:     7 0.006812177 0.53711117 0.3894486    Clic1       2      22
#> 17:     8 0.007043215 0.53711117 0.4882910   Tmbim1       2       6
#> 18:     9 0.007624607 0.53711117 0.5138513  Tmem88b       2       3
#> 19:    10 0.009028757 0.54975200 0.3764765    Cryab       2      26
#> 20:    12 0.011597115 0.54975200 0.4520588   Phldb1       2      10
#> 21:    13 0.012453659 0.54975200 0.4894421    Grb14       2       5
#> 22:    15 0.015239247 0.54975200 0.4697248 Ppp1r14a       2       8
#> 23:    27 0.028652205 0.54975200 0.5009713    Tnnt1       2       4
#> 24:    11 0.015565349 0.81477908 0.3794660     Bcam       3       4
#> 25:    20 0.026477390 0.81477908 0.4633258     Grm1       3       2
#> 26:    22 0.028812026 0.83031020 0.4163589    Itih3       3       3
#> 27:    55 0.091050217 0.96701893 0.3757642    Plcb4       3       5
#> 28:    61 0.100930232 0.96701893 0.4724775    Prkcd       3       1
#> 29:    64 0.110921899 0.96701893 0.3306064  Aldh1a1       3       7
#> 30:    65 0.111962853 0.96701893 0.3296060      Vim       3       8
#> 31:    66 0.114585832 0.96701893 0.2897821    Prox1       3      10
#> 32:    73 0.133643639 0.96701893 0.3633652     Zic1       3       6
#> 33:   107 0.175926132 0.96701893 0.3074804    Vamp1       3       9
#> 34:     7 0.003429921 0.28637427 0.6202705    Pcdh8       4       2
#> 35:     9 0.004989616 0.34768541 0.5338170      Gss       4       5
#> 36:    10 0.005483997 0.34768541 0.6350344     Cygb       4       1
#> 37:    11 0.008053972 0.41605624 0.5981233   Sowaha       4       3
#> 38:    25 0.021696009 0.52964105 0.4588221   Homer3       4       9
#> 39:    31 0.027383336 0.53917299 0.5491473    Wipf3       4       4
#> 40:    33 0.029708388 0.53917299 0.5052576    C1ql2       4       7
#> 41:    38 0.033496196 0.53917299 0.5205546    Calb1       4       6
#> 42:    45 0.038269376 0.53917299 0.4568997   Vstm2l       4      10
#> 43:    73 0.076839208 0.66734326 0.4669022    Synpr       4       8
#> 44:    10 0.032781719 0.96693210 0.4073136  Ppp1r1a       5       5
#> 45:    13 0.035476975 0.96693210 0.3952556 Serpinh1       5       7
#> 46:    16 0.039171568 0.96693210 0.4033382   Cacng5       5       6
#> 47:    17 0.041923751 0.96693210 0.3813645   Suclg2       5       9
#> 48:    23 0.050830947 0.96693210 0.4847041   Bcl11b       5       1
#> 49:    29 0.065093223 0.96693210 0.3870576    Grb14       5       8
#> 50:    35 0.070743235 0.96693210 0.4102664  Pitpnm2       5       4
#> 51:    40 0.091234404 0.96693210 0.4110672   Sowaha       5       3
#> 52:    60 0.118368025 0.96693210 0.4153359  Fam163b       5       2
#> 53:    64 0.121852917 0.96693210 0.3794176     Rora       5      10
#> 54:    13 0.001389098 0.06774522 0.8375194    Smim1       6       1
#> 55:    20 0.008407944 0.24881180 0.6904204      Ttr       6       7
#> 56:    25 0.010250009 0.24881180 0.8264599     Hopx       6       2
#> 57:    30 0.012423050 0.24976015 0.7589815  Siglech       6       5
#> 58:    33 0.013357161 0.24976015 0.7804225    Cldn5       6       3
#> 59:    39 0.022190202 0.36073304 0.6920161   Homer2       6       6
#> 60:    44 0.029494591 0.42499024 0.6463815    Evi2a       6       8
#> 61:    45 0.030635500 0.43137095 0.7611254    Cabp7       6       4
#> 62:    47 0.032523726 0.43137095 0.6352972     Vsir       6       9
#> 63:    51 0.036742690 0.45676206 0.6266049  Cyp2d22       6      10
#> 64:    77 0.019024942 0.15664692 1.0917235     Sypl       7       4
#> 65:    80 0.038112923 0.30204492 1.0720757     Vwa1       7       6
#> 66:    87 0.065087126 0.46778865 1.0877216    Man1a       7       5
#> 67:    92 0.075133648 0.51776883 1.1714598    Olig2       7       2
#> 68:   103 0.095480764 0.58771655 1.0561925     Gng4       7       7
#> 69:   104 0.096534827 0.58849116 1.1651781     Myrf       7       3
#> 70:   111 0.112835941 0.64448637 0.9399060   Pbxip1       7      10
#> 71:   113 0.125182233 0.70234987 1.2522299   Pou3f1       7       1
#> 72:   117 0.133698127 0.72448387 0.9716501     Aspa       7       9
#> 73:   127 0.179546145 0.89631698 1.0321280     Lmo1       7       8
#>       Top     p.value        FDR     logFC    feats cluster ranking