findScranMarkers_one_vs_all

Identify marker genes for all clusters in a one vs all manner based on scran's implementation of findMarkers.

findScranMarkers_one_vs_all(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  pval = 0.01,
  logFC = 0.5,
  min_genes = 10,
  verbose = TRUE,
  ...
)

Arguments

gobject
giotto object

expression_values
gene expression values to use

cluster_column
clusters to use

subset_clusters
subset of clusters to use

pval
filter on minimal p-value

logFC
filter on logFC

min_genes
minimum genes to keep per cluster, overrides pval and logFC

verbose
be verbose

...
additional parameters for the findMarkers function in scran

Value

data.table with marker genes

See also

Examples

data(mini_giotto_single_cell)

scran_markers = findScranMarkers_one_vs_all(gobject = mini_giotto_single_cell,
                                            cluster_column = 'leiden_clus')

#> using 'Scran' to detect marker genes. If used in published research, please cite:
#>   Lun ATL, McCarthy DJ, Marioni JC (2016).
#>   'A step-by-step workflow for low-level analysis of single-cell RNA-seq data with Bioconductor.'
#>   F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2. 

#> 
#>  start with cluster  1 
#> 
#>  start with cluster  2 
#> 
#>  start with cluster  3