makeSignMatrixDWLS

Function to convert a matrix within a Giotto object into a format that can be used with runDWLSDeconv for deconvolution. A vector of cell types for parameter cell_type_vector can be created from the cell metadata (pDataDT).

makeSignMatrixDWLS(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log = TRUE,
  log_base = 2,
  sign_gene,
  cell_type_vector,
  cell_type = NULL
)

Arguments

gobject

Giotto object of single cell

spat_unit

spatial unit

feat_type

feature type to use

expression_values

expression values to use

reverse_log

reverse a log-normalized expression matrix

log_base

the logarithm base (default = 2)

sign_gene

all of DE genes (signature)

cell_type_vector

vector with cell types (length = ncol(matrix))

cell_type

deprecated, use cell_type_vector

Value

matrix

See also

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used
sign_gene <- c("Bcl11b", "Lmo1", "F3", "Cnih3", "Ppp1r3c", "Rims2", "Gfap",
"Gjc3", "Chrna4", "Prkcd", "Prr18", "Grb14", "Tprn", "Clic1", "Olig2",
"Hrh3", "Tmbim1", "Carhsp1", "Tmem88b", "Ugt8a", "Arpp19", "Lamp5",
"Galnt6", "Hlf", "Hs3st2", "Tbr1", "Myl4", "Cygb", "Ttc9b","Ipcef1")

makeSignMatrixDWLS(gobject = g, sign_gene = sign_gene,
cell_type_vector = pDataDT(g)[["leiden_clus"]])
#>                 2         1         3         7          6         8         4
#> Hlf     15.051790 17.174010 18.022425 21.301625 13.2632813 22.019479 16.990281
#> Tmbim1   7.986970  6.632170  6.786006  7.343705  5.1000630  2.246642  5.960372
#> Clic1    2.485635  1.734531  1.775180  2.280386  1.8334849  2.214839  1.505498
#> Arpp19  35.437112 34.538454 40.161748 28.130800 26.3630445 18.450100 36.098097
#> Ttc9b   37.837127 38.713193 47.032313 39.430508 34.3414148 24.993843 43.358528
#> Carhsp1  8.705566  6.401236  9.165021 10.134379  7.4322546  6.264241  6.751019
#> Cygb     5.490290  4.361537  7.164488  6.376469  5.8536978  7.283729  5.529844
#> Gfap    18.938024 21.563875 18.343494 15.390794 17.5382407 17.242703 17.196169
#> Prkcd   28.778941 30.424482 31.825506 36.681868 22.9861436 46.273348 17.177341
#> Cnih3    4.641055  6.524519  7.489424  8.485436  2.8790300  1.168333  4.882899
#> Grb14   18.652498 14.503625 13.788083 15.994655 15.7385073  9.505730 10.581237
#> Lamp5    7.984634  9.265580 10.726291  7.386576  3.9270151  2.755634  9.571582
#> Chrna4   5.832506  7.434218  7.554707  6.646480  4.1787364  7.912088  5.760569
#> F3       3.636223  5.724523  3.128278  4.057015  4.5195901  4.001789  5.639305
#> Ugt8a   12.819532 11.081179 10.269863 11.284812 10.2048666  2.214839  8.610902
#> Tbr1    10.096214  9.544250 13.040371  9.126290  8.9160227  2.723832  9.384128
#> Lmo1     6.393886  8.731378  6.952993  9.571341  7.1269624  7.066768  9.218877
#> Galnt6   2.051510  1.399854  2.561461  1.320149  0.7115607  0.000000  1.847800
#> Rims2    8.159905  9.198734  7.784488  9.992842  7.9585286  4.583308  6.939767
#> Hrh3     8.893634  7.767136  9.214517  9.344162  5.8699575  7.375856  6.082125
#> Olig2    4.886196  3.938366  3.407480  4.539122  3.1845920  4.860594  3.579956
#> Hs3st2   4.039976  3.199028  5.515714  5.266052  3.3312723  2.279649  3.785530
#> Bcl11b   8.974808 12.888756  9.338095 11.154306  7.6240480 11.769898 11.020617
#> Tprn     4.912682  3.463219  4.363062  5.868955  4.0812924  7.375856  2.438520
#> Prr18   11.148036  8.380231  8.103621  6.275658  8.3796725 21.561416  5.865312
#> Gjc3     7.175789  8.065834  6.596657  5.987060  7.4479619  5.622096  6.463732
#> Myl4     5.131439  6.564998  8.875452  6.189511  3.8530311  5.129214  5.694683
#> Ipcef1   6.265445  6.455509  8.241332  5.951944  5.1048947  3.900256  6.646003
#> Ppp1r3c  3.752126  4.225897  3.142713  2.648930  3.0046784  3.207609  2.475447
#> Tmem88b 14.087201 11.557085 12.512200 12.596784 11.0238146 12.217616  8.893961
#>                  5
#> Hlf     14.3469314
#> Tmbim1   7.1898449
#> Clic1    1.3758725
#> Arpp19  31.6774100
#> Ttc9b   35.6106358
#> Carhsp1  5.9259790
#> Cygb     9.9457990
#> Gfap    15.5332726
#> Prkcd   31.5894559
#> Cnih3    7.1019349
#> Grb14   13.8366602
#> Lamp5    6.4794106
#> Chrna4   6.8132600
#> F3       3.5376041
#> Ugt8a   10.9697049
#> Tbr1     8.1725361
#> Lmo1     9.4419770
#> Galnt6   0.8065613
#> Rims2    7.9002783
#> Hrh3     6.9449353
#> Olig2    3.5420465
#> Hs3st2   2.5846131
#> Bcl11b   9.2230781
#> Tprn     3.0822527
#> Prr18   10.7006862
#> Gjc3     7.1461107
#> Myl4     4.3237500
#> Ipcef1   4.8603412
#> Ppp1r3c  4.5186362
#> Tmem88b 12.0655801