makeSignMatrixRank

Function to convert a single-cell count matrix and a corresponding single-cell cluster vector into a rank matrix that can be used with the Rank enrichment option.

makeSignMatrixRank(
  sc_matrix,
  sc_cluster_ids,
  ties_method = c("random", "max"),
  gobject = NULL
)

Arguments

sc_matrix

matrix of single-cell RNAseq expression data

sc_cluster_ids

vector of cluster ids

ties_method

how to handle rank ties

gobject

if giotto object is given then only genes present in both datasets will be considered

Value

matrix

See also

Examples

sign_gene <- c(
    "Bcl11b", "Lmo1", "F3", "Cnih3", "Ppp1r3c", "Rims2", "Gfap",
    "Gjc3", "Chrna4", "Prkcd", "Prr18", "Grb14", "Tprn", "Clic1", "Olig2",
    "Hrh3", "Tmbim1", "Carhsp1", "Tmem88b", "Ugt8a", "Arpp19", "Lamp5",
    "Galnt6", "Hlf", "Hs3st2", "Tbr1", "Myl4", "Cygb", "Ttc9b", "Ipcef1"
)

sign_matrix <- matrix(rnorm(length(sign_gene) * 3), nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- c("cell_type1", "cell_type2", "cell_type3")

makeSignMatrixRank(
    sc_matrix = sign_matrix,
    sc_cluster_ids = c("cell_type1", "cell_type2", "cell_type3")
)
#> Warning: NaNs produced
#> Warning: NaNs produced
#> 30 x 3 Matrix of class "dgeMatrix"
#>         cell_type1 cell_type2 cell_type3
#> Bcl11b          17          9         17
#> Lmo1            30         23          2
#> F3               4         20         18
#> Cnih3           25          2          9
#> Ppp1r3c          2         26         15
#> Rims2           26          6          5
#> Gfap            27          1         25
#> Gjc3             7         15         22
#> Chrna4           3         21         12
#> Prkcd            8         22          8
#> Prr18           18         19          3
#> Grb14           28         25         27
#> Tprn             5         29          6
#> Clic1           29         24         29
#> Olig2           21          7         16
#> Hrh3            16         11         14
#> Tmbim1          13          5         20
#> Carhsp1         22          4         11
#> Tmem88b         20         17          7
#> Ugt8a           14         16         10
#> Arpp19          12          8         21
#> Lamp5           19          3         19
#> Galnt6          11         13         13
#> Hlf             23         28         28
#> Hs3st2          24         30         30
#> Tbr1             6         14         24
#> Myl4             9         10         26
#> Cygb            15         18          4
#> Ttc9b           10         12         23
#> Ipcef1           1         27          1