runPCAprojection

runs a Principal Component Analysis on a random subet + projection

runPCAprojection(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  random_subset = 500,
  name = "pca.projection",
  feats_to_use = "hvf",
  return_gobject = TRUE,
  center = TRUE,
  scale_unit = TRUE,
  ncp = 100,
  method = c("irlba"),
  method_params = BiocParallel::SerialParam(),
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE,
  ...
)

Arguments

gobject

giotto object

spat_unit

spatial unit

feat_type

feature type

expression_values

expression values to use

reduction

cells or genes

random_subset

numeric. random subset to perform PCA on

name

arbitrary name for PCA run

feats_to_use

subset of features to use for PCA

return_gobject

boolean: return giotto object (default = TRUE)

center

center data first (default = TRUE)

scale_unit

scale features before PCA (default = TRUE)

ncp

number of principal components to calculate

method

which implementation to use

method_params

BiocParallelParam object

rev

do a reverse PCA

set_seed

use of seed

seed_number

seed number to use

verbose

verbosity of the function

...

additional parameters for PCA (see details)

Value

giotto object with updated PCA dimension recuction

Details

See runPCA and PCA for more information about other parameters. This PCA implementation is similar to runPCA, except that it performs PCA on a subset of the cells or features, and predict on the others. This can significantly increase speed without sacrificing accuracy too much.

feats_to_use = NULL: will use all features from the selected matrix

feats_to_use = <hvg name>: can be used to select a column name of highly variable features, created by (see calculateHVF)

feats_to_use = c('geneA', 'geneB', ...): will use all manually provided features