spatCellCellcom

Spatial Cell-Cell communication scores based on spatial expression of interacting cells

spatCellCellcom(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  spatial_network_name = "Delaunay_network",
  cluster_column = "cell_types",
  random_iter = 1000,
  feat_set_1,
  feat_set_2,
  gene_set_1 = NULL,
  gene_set_2 = NULL,
  log2FC_addendum = 0.1,
  min_observations = 2,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("feats", "cells"),
  do_parallel = TRUE,
  cores = NA,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = c("a little", "a lot", "none")
)

Arguments

gobject

giotto object to use

feat_type

feature type

spat_unit

spatial unit

spatial_network_name

spatial network to use for identifying interacting cells

cluster_column

cluster column with cell type information

random_iter

number of iterations

feat_set_1

first specific feature set from feature pairs

feat_set_2

second specific feature set from feature pairs

gene_set_1

deprecated, use feat_set_1

gene_set_2

deprecated, use feat_set_2

log2FC_addendum

addendum to add when calculating log2FC

min_observations

minimum number of interactions needed to be considered

detailed

provide more detailed information (random variance and z-score)

adjust_method

which method to adjust p-values

adjust_target

adjust multiple hypotheses at the cell or feature level

do_parallel

run calculations in parallel with mclapply

cores

number of cores to use if do_parallel = TRUE

set_seed

set a seed for reproducibility

seed_number

seed number

verbose

verbose

Value

Cell-Cell communication scores for feature pairs based on spatial interaction

Details

Statistical framework to identify if pairs of genes (such as ligand-receptor combinations) are expressed at higher levels than expected based on a reshuffled null distribution of feature expression values in cells that are spatially in proximity to each other.

LR_comb: Pair of ligand and receptor

lig_cell_type: cell type to assess expression level of ligand

lig_expr: average expression of ligand in lig_cell_type

ligand: ligand name

rec_cell_type: cell type to assess expression level of receptor

rec_expr: average expression of receptor in rec_cell_type

receptor: receptor name

LR_expr: combined average ligand and receptor expression

lig_nr: total number of cells from lig_cell_type that spatially interact with cells from rec_cell_type

rec_nr: total number of cells from rec_cell_type that spatially interact with cells from lig_cell_type

rand_expr: average combined ligand and receptor expression from random spatial permutations

av_diff: average difference between LR_expr and rand_expr over all random spatial permutations

sd_diff: (optional) standard deviation of the difference between LR_expr and rand_expr over all random spatial permutations

z_score: (optional) z-score

log2fc: log2 fold-change (LR_expr/rand_expr)

pvalue: p-value

LR_cell_comb: cell type pair combination

p.adj: adjusted p-value

PI: significance score: log2fc * -log10(p.adj)

Examples

g <- GiottoData::loadGiottoMini("visium")
#> 1. read Giotto object
#> 2. read Giotto feature information
#> 3. read Giotto spatial information
#> 3.1 read Giotto spatial shape information
#> cell_spatInfo_spatVector.shp
#> cell
#> 
#> 3.2 read Giotto spatial centroid information
#> cell
#> 
#> 3.3 read Giotto spatial overlap information
#> No overlaps were found, overlap loading will be
#>  skipped
#> 
#> 4. read Giotto image information
#> 
#> no external python path was provided,
#>  but a giotto python environment was found and
#>  will be used

spatCellCellcom(
    gobject = g,
    cluster_column = "leiden_clus",
    feat_set_1 = "Gm19935",
    feat_set_2 = "9630013A20Rik",
    verbose = "a lot",
    random_iter = 10
)
#> simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10simulation  1simulation  2simulation  3simulation  4simulation  5simulation  6simulation  7simulation  8simulation  9simulation  10
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#>                   LR_comb lig_cell_type   lig_expr  ligand rec_cell_type
#>                    <char>        <fctr>      <num>  <char>        <fctr>
#>  1: Gm19935-9630013A20Rik             8 4.42689142 Gm19935             8
#>  2: Gm19935-9630013A20Rik             8 3.53150263 Gm19935             2
#>  3: Gm19935-9630013A20Rik             2 2.79311868 Gm19935             8
#>  4: Gm19935-9630013A20Rik             8 2.56424083 Gm19935             1
#>  5: Gm19935-9630013A20Rik             1 1.93973137 Gm19935             8
#>  6: Gm19935-9630013A20Rik             3 0.52939188 Gm19935             4
#>  7: Gm19935-9630013A20Rik             1 1.38783642 Gm19935             5
#>  8: Gm19935-9630013A20Rik             1 0.32805003 Gm19935             2
#>  9: Gm19935-9630013A20Rik             1 0.37564508 Gm19935             1
#> 10: Gm19935-9630013A20Rik             1 0.21566558 Gm19935             4
#> 11: Gm19935-9630013A20Rik             6 0.74313660 Gm19935             3
#> 12: Gm19935-9630013A20Rik             1 0.38356330 Gm19935             6
#> 13: Gm19935-9630013A20Rik             3 0.33105646 Gm19935             1
#> 14: Gm19935-9630013A20Rik             7 0.51613367 Gm19935             6
#> 15: Gm19935-9630013A20Rik             1 0.26016513 Gm19935             3
#> 16: Gm19935-9630013A20Rik             2 0.12091613 Gm19935             1
#> 17: Gm19935-9630013A20Rik             6 0.18874411 Gm19935             1
#> 18: Gm19935-9630013A20Rik             3 0.46288825 Gm19935             2
#> 19: Gm19935-9630013A20Rik             4 0.14090804 Gm19935             1
#> 20: Gm19935-9630013A20Rik             2 0.00000000 Gm19935             3
#> 21: Gm19935-9630013A20Rik             6 0.37485392 Gm19935             6
#> 22: Gm19935-9630013A20Rik             3 0.29024782 Gm19935             3
#> 23: Gm19935-9630013A20Rik             2 0.10979089 Gm19935             2
#> 24: Gm19935-9630013A20Rik             4 0.19385173 Gm19935             4
#> 25: Gm19935-9630013A20Rik             6 0.16830819 Gm19935             4
#> 26: Gm19935-9630013A20Rik             7 0.18697910 Gm19935             3
#> 27: Gm19935-9630013A20Rik             3 0.14993201 Gm19935             7
#> 28: Gm19935-9630013A20Rik             3 0.22663666 Gm19935             6
#> 29: Gm19935-9630013A20Rik             5 0.11014816 Gm19935             4
#> 30: Gm19935-9630013A20Rik             4 0.13249400 Gm19935             5
#> 31: Gm19935-9630013A20Rik             7 0.08578059 Gm19935             7
#> 32: Gm19935-9630013A20Rik             5 0.10179997 Gm19935             5
#> 33: Gm19935-9630013A20Rik             7 0.08602228 Gm19935             1
#> 34: Gm19935-9630013A20Rik             1 0.00000000 Gm19935             7
#> 35: Gm19935-9630013A20Rik             5 0.00000000 Gm19935             1
#> 36: Gm19935-9630013A20Rik             4 0.00000000 Gm19935             3
#> 37: Gm19935-9630013A20Rik             6 0.00000000 Gm19935             7
#> 38: Gm19935-9630013A20Rik             4 0.00000000 Gm19935             6
#>                   LR_comb lig_cell_type   lig_expr  ligand rec_cell_type
#>       rec_expr      receptor    LR_expr lig_nr rec_nr  rand_expr      av_diff
#>          <num>        <char>      <num>  <int>  <int>      <num>        <num>
#>  1: 0.36930096 9630013A20Rik 4.79619238      5      5 4.83203241 -0.035840033
#>  2: 0.85113081 9630013A20Rik 4.38263344      3      5 2.79969356  1.582939880
#>  3: 0.36930096 9630013A20Rik 3.16241964      5      3 3.28428278 -0.121863137
#>  4: 0.32932035 9630013A20Rik 2.89356118      7     17 2.01223816  0.881323023
#>  5: 0.25386055 9630013A20Rik 2.19359192     17      7 2.27077367 -0.077181743
#>  6: 1.27138464 9630013A20Rik 1.80077651      5      3 0.46719218  1.333584337
#>  7: 0.00000000 9630013A20Rik 1.38783642      3      2 0.27499691  1.112839515
#>  8: 0.60854018 9630013A20Rik 0.93659021     50     39 0.63911857  0.297471638
#>  9: 0.55406806 9630013A20Rik 0.92971315    143    143 0.88781052  0.041902629
#> 10: 0.61326784 9630013A20Rik 0.82893342     17     15 0.75273218  0.076201236
#> 11: 0.00000000 9630013A20Rik 0.74313660     15     25 0.26737714  0.475759463
#> 12: 0.22594046 9630013A20Rik 0.60950376     42     42 0.70898665 -0.099482897
#> 13: 0.24285335 9630013A20Rik 0.57390981     77     82 0.66662904 -0.092719228
#> 14: 0.00000000 9630013A20Rik 0.51613367      3      2 0.33858193  0.177551740
#> 15: 0.25414463 9630013A20Rik 0.51430977     82     77 0.68987764 -0.175567878
#> 16: 0.35996626 9630013A20Rik 0.48088239     39     50 0.62149701 -0.140614626
#> 17: 0.28074309 9630013A20Rik 0.46948720     42     42 0.62040965 -0.150922445
#> 18: 0.00000000 9630013A20Rik 0.46288825      8      6 0.42523635  0.037651899
#> 19: 0.32184532 9630013A20Rik 0.46275337     15     17 0.55599367 -0.093240307
#> 20: 0.46057019 9630013A20Rik 0.46057019      6      8 0.43002647  0.030543722
#> 21: 0.07387806 9630013A20Rik 0.44873198     54     54 0.41971927  0.029012707
#> 22: 0.12322469 9630013A20Rik 0.41347251     87     87 0.42300381 -0.009531296
#> 23: 0.29867022 9630013A20Rik 0.40846111    125    125 0.42565825 -0.017197139
#> 24: 0.19923762 9630013A20Rik 0.39308935     71     71 0.37946575  0.013623598
#> 25: 0.12244051 9630013A20Rik 0.29074870     13     12 0.38363528 -0.092886581
#> 26: 0.08370451 9630013A20Rik 0.27068361     43     47 0.37140287 -0.100719258
#> 27: 0.08303690 9630013A20Rik 0.23296890     47     43 0.28667466 -0.053705752
#> 28: 0.00000000 9630013A20Rik 0.22663666     25     15 0.53757677 -0.310940108
#> 29: 0.11599635 9630013A20Rik 0.22614451     60     53 0.23235890 -0.006214385
#> 30: 0.05329353 9630013A20Rik 0.18578753     53     60 0.27360655 -0.087819018
#> 31: 0.08303690 9630013A20Rik 0.16881749     43     43 0.16881749  0.000000000
#> 32: 0.04844866 9630013A20Rik 0.15024863     66     66 0.15024863  0.000000000
#> 33: 0.00000000 9630013A20Rik 0.08602228     18     14 0.61614452 -0.530122241
#> 34: 0.07941508 9630013A20Rik 0.07941508     14     18 0.55532931 -0.475914228
#> 35: 0.00000000 9630013A20Rik 0.00000000      2      3 0.41746934 -0.417469345
#> 36: 0.00000000 9630013A20Rik 0.00000000      3      5 0.06820362 -0.068203616
#> 37: 0.00000000 9630013A20Rik 0.00000000      2      3 0.43604006 -0.436040064
#> 38: 0.00000000 9630013A20Rik 0.00000000     12     13 0.20825921 -0.208259206
#>       rec_expr      receptor    LR_expr lig_nr rec_nr  rand_expr      av_diff
#>            log2fc pvalue LR_cell_comb p.adj           PI
#>             <num>  <num>       <char> <num>        <num>
#>  1: -1.052204e-02    0.4         8--8   0.4 -0.004187139
#>  2:  6.284461e-01    0.0         8--2   0.0  0.060902718
#>  3: -5.290775e-02    0.8         2--8   0.8 -0.005127291
#>  4:  5.030902e-01    0.0         8--1   0.0  0.000000000
#>  5: -4.774921e-02    1.0         1--8   1.0  0.000000000
#>  6:  1.744679e+00    0.0         3--4   0.0  0.079832150
#>  7:  1.988265e+00    0.2         1--5   0.2  1.389737816
#>  8:  4.879679e-01    0.2         1--2   0.2  0.341074957
#>  9:  5.993626e-02    0.0         1--1   0.0         -Inf
#> 10:  1.234825e-01    0.6         1--4   0.6  0.027394436
#> 11:  1.198505e+00    0.2         6--3   0.2  0.837718719
#> 12: -1.893056e-01    0.6         1--6   0.6 -0.041997206
#> 13: -1.859731e-01    0.8         3--1   0.8 -0.018022657
#> 14:  4.903970e-01    1.0         7--6   1.0  0.000000000
#> 15: -3.626629e-01    0.2         1--3   0.2 -0.253490469
#> 16: -3.127473e-01    0.6         2--1   0.6 -0.069382605
#> 17: -3.391541e-01    0.4         6--1   0.4 -0.134962977
#> 18:  9.988177e-02    0.8         3--2   0.8  0.009679543
#> 19: -2.211791e-01    1.0         4--1   1.0  0.000000000
#> 20:  8.083062e-02    1.0         2--3   1.0  0.000000000
#> 21:  7.836910e-02    0.5         6--6   0.5  0.023591450
#> 22: -2.653440e-02    0.8         3--3   0.8 -0.002571449
#> 23: -4.798772e-02    0.4         2--2   0.4 -0.019096234
#> 24:  4.042133e-02    0.7         4--4   0.7  0.006261343
#> 25: -3.076784e-01    0.6         6--4   0.6 -0.068258072
#> 26: -3.467722e-01    0.6         7--3   0.6 -0.076930981
#> 27: -2.157328e-01    0.4         3--7   0.4 -0.085848695
#> 28: -9.649123e-01    0.0         3--6   0.0 -0.674444784
#> 29: -2.723062e-02    1.0         5--4   1.0  0.000000000
#> 30: -3.865768e-01    0.2         4--5   0.2 -0.270205565
#> 31:  0.000000e+00    1.0         7--7   1.0  0.000000000
#> 32: -3.203427e-16    1.0         5--5   1.0  0.000000000
#> 33: -1.944775e+00    0.0         7--1   0.0          Inf
#> 34: -1.868919e+00    0.0         1--7   0.0          Inf
#> 35: -2.371473e+00    0.7         5--1   0.7 -0.367345878
#> 36: -7.502087e-01    0.9         4--3   0.9 -0.034327668
#> 37: -2.422341e+00    0.7         6--7   0.7 -0.375225343
#> 38: -1.624144e+00    0.2         4--6   0.2 -1.135227924
#>            log2fc pvalue LR_cell_comb p.adj           PI